Bug fixes for ext_proc - porting from internal repo.
