+++ orion-lib/src/listeners/http_connection_manager/ext_proc.rs
@@ -20,11 +20,17 @@ mod mutation;
 mod r#override;
 mod processing;
 mod status;
+#[cfg(test)]
+mod tests;
 mod worker_config;
 
-use crate::body::channel_body::{ChannelBody, FrameBridge};
-
+use crate::body::{
+    body_with_metrics::BodyWithMetrics,
+    channel_body::{ChannelBody, FrameBridge},
+};
 use crate::event_error::EventKind;
+use http_body_util::{BodyExt, Collected, LengthLimitError, Limited};
+
 use crate::listeners::http_connection_manager::ext_proc::mutation::{
     apply_request_header_mutations, apply_response_header_mutations,
 };
@@ -36,7 +42,7 @@ use crate::listeners::http_connection_manager::ext_proc::status::ProcessingStatu
 use crate::listeners::http_connection_manager::ext_proc::status::ReadyStatus;
 use crate::listeners::http_connection_manager::ext_proc::worker_config::ExternalProcessingWorkerConfig;
 use crate::{
-    body::{body_with_metrics::BodyWithMetrics, response_flags::ResponseFlags},
+    body::response_flags::ResponseFlags,
     clusters::clusters_manager::{self, RoutingContext},
     listeners::{http_connection_manager::FilterDecision, synthetic_http_response::SyntheticHttpResponse},
     Error, PolyBody,
@@ -44,11 +50,8 @@ use crate::{
 use bytes::Bytes;
 use futures::{future::Either, StreamExt};
 use http::header::CONTENT_LENGTH;
-use http::{Request, Response};
-use http_body::Frame;
-use http_body_util::combinators::WithTrailers;
-use http_body_util::BodyExt;
-use http_body_util::Collected;
+use http::{Request, Response, StatusCode};
+use http_body::{Body, Frame};
 use http_body_util::Full;
 use orion_configuration::config::{
     cluster::ClusterSpecifier,
@@ -75,20 +78,27 @@ use orion_format::types::ResponseFlags as FmtResponseFlags;
 use pingora_timeout::fast_timeout;
 use scopeguard::defer;
 use std::convert::Infallible;
-use std::future::Future;
+use std::future::ready;
 use std::{sync::Arc, time::Duration};
 use tokio::sync::mpsc::error::SendError;
 use tokio::sync::{mpsc, oneshot};
-use tracing::{debug, error, warn};
+use tracing::{debug, error, info, warn};
+
+const EXT_PROC_BUFFERED_BODY_LIMIT: usize = 4 * 1024 * 1024; // this is the default limit for GRPC payload lenght
 
 #[derive(Debug, Clone, thiserror::Error)]
 pub enum ExtProcError {
     #[error("Timeout Error: {0}")]
     Timeout(&'static str),
+    #[error("Unsupported Response Type: {0}")]
+    UnsupportedResponseType(String),
+    #[error("Connection closed by remote GRPC")]
+    UnexpectedEof,
+    #[error("GRPC error: {0}")]
+    GrpcError(String),
 }
 
 #[derive(Debug, Clone)]
-#[allow(clippy::struct_excessive_bools)]
 pub struct ExternalProcessor {
     ext_proc_worker: Option<mpsc::Sender<ProcessingTask>>,
     worker_config: Arc<ExternalProcessingWorkerConfig>,
@@ -105,7 +115,7 @@ impl From<ExternalProcessorConfig> for ExternalProcessor {
 
 impl From<(ExternalProcessorConfig, Option<ExtProcPerRoute>)> for ExternalProcessor {
     fn from((initial_config, per_route_config): (ExternalProcessorConfig, Option<ExtProcPerRoute>)) -> Self {
-        debug!(target: "ext_proc", "From<ExternalProcessorConfig, ExtProcPerRoute> for ExternalProcessor");
+        debug!(target: "ext_proc", "From<ExternalProcessorConfig, Option<ExtProcPerRoute>> for ExternalProcessor");
         let forward_rules = initial_config.forward_rules.clone().map(Arc::new);
         let worker_config = ExternalProcessingWorkerConfig::from((initial_config, per_route_config));
 
@@ -173,13 +183,14 @@ macro_rules! run_action {
     // $processor: The specific processing struct (e.g., self.request_processing or self.response_processing).
     // $action: The Action to process.
     // $ctx: A string literal context for logging.
-    ($self:ident, $processor:expr, $action:expr, $ctx:literal) => {
+    ($self:ident, $processor:expr, $action:expr, $ctx:expr) => {
         match $action {
             Action::Send(outbound) => {
                 debug!(target: "ext_proc", "{ctx} @{typ}: action -> forward {outbound:?}",
                     ctx = $ctx,
                     typ = stringify!($processor),
-                    outbound = outbound);
+                    outbound = outbound
+                );
 
                 $self.forward_to_external_processor(outbound).await;
             },
@@ -222,14 +233,10 @@ macro_rules! with_current_processing {
 }
 
 impl ExternalProcessor {
-    fn collect_to_single_chunk(
-        original: Collected<Bytes>,
-    ) -> WithTrailers<Full<Bytes>, impl Future<Output = Option<Result<http::HeaderMap, Infallible>>>> {
-        let trailers = original.trailers().cloned();
+    async fn to_buffered(original: Collected<Bytes>) -> Collected<Bytes> {
+        let trailers = original.trailers().cloned().map(Ok::<_, Infallible>);
         let aggregated_bytes = original.to_bytes();
-        let new_body = Full::new(aggregated_bytes);
-        let trailer_future = async move { trailers.map(Ok::<_, Infallible>) };
-        new_body.with_trailers(trailer_future)
+        Full::new(aggregated_bytes).with_trailers(ready(trailers)).collect().await.unwrap()
     }
 
     #[allow(clippy::too_many_lines)]
@@ -256,32 +263,50 @@ impl ExternalProcessor {
             (OverridableBodyMode::None, trailers_mode) => {
                 // event though body processing is None and trailers processing is Skip, we have to
                 // create the bridge, to allow ext_proc mutate the body with ContinueAndReplace action.
-                debug!(target: "ext_proc", "request processing body(None) and trailers:{trailers_mode:?} => {body:?}");
+                debug!(target: "ext_proc", "request processing body(None) and trailers:{trailers_mode:?}");
                 let (new_body, bridge) = ChannelBody::new(body);
                 request.body_mut().inner = PolyBody::from(new_body);
                 bridge
             },
             (OverridableBodyMode::Streamed | OverridableBodyMode::FullDuplexStreamed, trailers_mode) => {
-                debug!(target: "ext_proc", "request processing body(Streamed) and trailers:{trailers_mode:?} => {body:?}");
+                debug!(target: "ext_proc", "request processing body(Streamed) and trailers:{trailers_mode:?}");
                 let (new_body, bridge) = ChannelBody::new(body);
                 request.body_mut().inner = PolyBody::from(new_body);
                 bridge
             },
             (OverridableBodyMode::Buffered | OverridableBodyMode::BufferedPartial, trailers_mode) => {
-                debug!(target: "ext_proc", "request processing body(Buffered) with trailers:{trailers_mode:?} => {body:?}");
-                let Ok(collected) = body.collect().await else {
-                    return self.on_filter_error(
-                        "Failed to collect request body for external processing",
-                        None,
-                        request.version(),
-                    );
-                };
+                debug!(target: "ext_proc", "request processing body(Buffered) with trailers:{trailers_mode:?}");
+                if body.is_end_stream() {
+                    let (new_body, bridge) = ChannelBody::new(body);
+                    request.body_mut().inner = PolyBody::from(new_body);
+                    bridge
+                } else {
+                    let body = Limited::new(body, EXT_PROC_BUFFERED_BODY_LIMIT);
+                    let collected = match body.collect().await {
+                        Ok(collected) => collected,
+                        Err(e) => {
+                            if let Some(_) = e.downcast_ref::<LengthLimitError>() {
+                                return self.on_filter_error(
+                                    &format!("Request body: {}", e),
+                                    None,
+                                    request.version(),
+                                    Some(StatusCode::PAYLOAD_TOO_LARGE),
+                                );
+                            }
+                            return self.on_filter_error(
+                                &format!("Error collecting request body: {}", e),
+                                None,
+                                request.version(),
+                                None,
+                            );
+                        },
+                    };
 
-                let buffered = Self::collect_to_single_chunk(collected);
-
-                let (new_body, bridge) = ChannelBody::new(buffered);
-                request.body_mut().inner = PolyBody::from(new_body);
-                bridge
+                    let buffered = Self::to_buffered(collected).await;
+                    let (new_body, bridge) = ChannelBody::new(buffered);
+                    request.body_mut().inner = PolyBody::from(new_body);
+                    bridge
+                }
             },
         };
 
@@ -292,7 +317,12 @@ impl ExternalProcessor {
 
         let ver = request.version();
         let Ok(response_rx) = self.send_processing_data(processing_data, ver).await else {
-            return self.on_filter_error("Failed to schedule sending request data to external processor", None, ver);
+            return self.on_filter_error(
+                "Failed to schedule sending request data to external processor",
+                None,
+                ver,
+                None,
+            );
         };
 
         let res = match response_rx.await {
@@ -317,6 +347,7 @@ impl ExternalProcessor {
                             "Invalid header modifications received from external processor",
                             Some(e),
                             request.version(),
+                            None,
                         );
                     }
                 }
@@ -332,16 +363,20 @@ impl ExternalProcessor {
                 FilterDecision::Continue
             },
             Ok(ProcessingStatus::ResponseReady(ReadyStatus { .. })) => {
-                warn!(target: "ext_proc", "apply_request: unexpected ResponseReady...");
+                warn!(target: "ext_proc", "apply_request: unexpected ResponseReady!");
                 return self.on_filter_error(
                     "Unexpected ResponseReady status received during request processing",
                     None,
                     request.version(),
+                    None,
                 );
             },
-            Err(e) => {
-                self.on_filter_error(format!("External processor: {e:?}").as_str(), Some(e.into()), request.version())
-            },
+            Err(e) => self.on_filter_error(
+                format!("External processor: {e:?}").as_str(),
+                Some(e.into()),
+                request.version(),
+                None,
+            ),
         };
 
         debug!(target: "ext_proc", "apply_request completed: {res:?}!");
@@ -367,38 +402,56 @@ impl ExternalProcessor {
             ext_proc_headers = Some(self.filter_header_map(response.headers()));
         }
 
-        let body: PolyBody = std::mem::take(response.body_mut());
+        let body: PolyBody = std::mem::take(&mut response.body_mut());
 
         let ext_proc_frame_bridge = match (modes.body_mode(), modes.trailer_mode()) {
             (OverridableBodyMode::None, trailers_mode) => {
                 // event though body processing is None and trailers processing is Skip, we have to
                 // create the bridge, to allow ext_proc mutate the body with ContinueAndReplace action.
-                debug!(target: "ext_proc", "response processing body(None) and trailers:{trailers_mode:?} => {body:?}");
+                debug!(target: "ext_proc", "response processing body(None) and trailers:{trailers_mode:?}");
                 let (new_body, bridge) = ChannelBody::new(body);
                 *response.body_mut() = PolyBody::from(new_body);
                 bridge
             },
             (OverridableBodyMode::Streamed | OverridableBodyMode::FullDuplexStreamed, trailers_mode) => {
-                debug!(target: "ext_proc", "response processing body(Streamed) and trailers:{trailers_mode:?} => {body:?}");
+                debug!(target: "ext_proc", "response processing body(Streamed) and trailers:{trailers_mode:?}");
                 let (new_body, bridge) = ChannelBody::new(body);
                 *response.body_mut() = PolyBody::from(new_body);
                 bridge
             },
             (OverridableBodyMode::Buffered | OverridableBodyMode::BufferedPartial, trailers_mode) => {
-                debug!(target: "ext_proc", "response processing body(Buffered) with trailers:{trailers_mode:?} => {body:?}");
-                let Ok(collected) = body.collect().await else {
-                    return self.on_filter_error(
-                        "Failed to collect response body for external processing",
-                        None,
-                        response.version(),
-                    );
-                };
+                debug!(target: "ext_proc", "response processing body(Buffered) with trailers:{trailers_mode:?}");
+                if body.is_end_stream() {
+                    let (new_body, bridge) = ChannelBody::new(body);
+                    *response.body_mut() = PolyBody::from(new_body);
+                    bridge
+                } else {
+                    let body = Limited::new(body, EXT_PROC_BUFFERED_BODY_LIMIT);
+                    let collected = match body.collect().await {
+                        Ok(collected) => collected,
+                        Err(e) => {
+                            if let Some(_) = e.downcast_ref::<LengthLimitError>() {
+                                return self.on_filter_error(
+                                    &format!("Response body: {}", e),
+                                    None,
+                                    response.version(),
+                                    Some(StatusCode::PAYLOAD_TOO_LARGE),
+                                );
+                            }
+                            return self.on_filter_error(
+                                &format!("Error collecting response body: {}", e),
+                                None,
+                                response.version(),
+                                None,
+                            );
+                        },
+                    };
 
-                debug!(target: "ext_proc", "response body collected: {collected:?}");
-                let buffered = Self::collect_to_single_chunk(collected);
-                let (new_body, bridge) = ChannelBody::new(buffered);
-                *response.body_mut() = PolyBody::from(new_body);
-                bridge
+                    let buffered = Self::to_buffered(collected).await;
+                    let (new_body, bridge) = ChannelBody::new(buffered);
+                    *response.body_mut() = PolyBody::from(new_body);
+                    bridge
+                }
             },
         };
 
@@ -409,7 +462,12 @@ impl ExternalProcessor {
 
         let ver = response.version();
         let Ok(response_rx) = self.send_processing_data(processing_data, ver).await else {
-            return self.on_filter_error("Failed to schedule sending response data to external processor", None, ver);
+            return self.on_filter_error(
+                "Failed to schedule sending response data to external processor",
+                None,
+                ver,
+                None,
+            );
         };
 
         let res = match response_rx.await {
@@ -434,6 +492,7 @@ impl ExternalProcessor {
                             "Invalid header modifications received from external processor",
                             Some(e),
                             response.version(),
+                            None,
                         );
                     }
                 }
@@ -446,17 +505,19 @@ impl ExternalProcessor {
                 FilterDecision::Continue
             },
             Ok(ProcessingStatus::RequestReady(ReadyStatus { .. })) => {
-                warn!(target: "ext_proc", "apply_response: unexpected RequestReady...");
+                warn!(target: "ext_proc", "apply_response: unexpected RequestReady!");
                 return self.on_filter_error(
                     "Unexpected RequestReady status received during response processing",
                     None,
                     response.version(),
+                    None,
                 );
             },
             Err(e) => self.on_filter_error(
                 format!("External processor response processing: {e:?}").as_str(),
                 Some(e.into()),
                 response.version(),
+                None,
             ),
         };
 
@@ -485,19 +546,25 @@ impl ExternalProcessor {
         worker_channel.send(processing_message).await.map(|()| response_rx)
     }
 
-    fn on_filter_error(&mut self, msg: &str, error: Option<Error>, http_version: http::Version) -> FilterDecision {
+    fn on_filter_error(
+        &mut self,
+        msg: &str,
+        error: Option<Error>,
+        http_version: http::Version,
+        status_code: Option<StatusCode>,
+    ) -> FilterDecision {
         if let Some(err) = error {
-            warn!(target: "ext_proc","{msg}: {err}");
+            info!(target: "ext_proc","{msg}: {err}");
         } else {
-            warn!(target: "ext_proc", "{msg}");
+            info!(target: "ext_proc", "{msg}");
         }
         if self.worker_config.failure_mode_allow {
             FilterDecision::Continue
         } else {
             FilterDecision::DirectResponse(
-                SyntheticHttpResponse::internal_error_with_msg(
-                    msg,
-                    EventKind::FilterChainNotFound.into(),
+                SyntheticHttpResponse::custom_error(
+                    status_code.unwrap_or(StatusCode::INTERNAL_SERVER_ERROR),
+                    EventKind::ExtProcError.into(),
                     ResponseFlags(FmtResponseFlags::UPSTREAM_CONNECTION_FAILURE),
                 )
                 .into_response(http_version),
@@ -658,6 +725,13 @@ fn clone_frame(frame: &Frame<Bytes>) -> Frame<Bytes> {
     }
 }
 
+#[allow(dead_code)]
+enum MergeResult {
+    Retry(u32),
+    Error(u32),
+    None(u32),
+}
+
 impl ExternalProcessingWorker<kind::Processing> {
     fn new(config: Arc<ExternalProcessingWorkerConfig>, overridable_global_modes: Arc<OverridableGlobalModes>) -> Self {
         let request_processing = RequestProcessing::<kind::Processing>::from(&*config);
@@ -676,6 +750,39 @@ impl ExternalProcessingWorker<kind::Processing> {
             handshake,
             timeout_state: TimeoutState { active: false, duration: message_timeout, extended: false },
             overridable_modes: overridable_global_modes,
+        }
+    }
+
+    async fn recover_or_failure(&mut self, err: ExtProcError, log_msg: &str) {
+        if self.config.failure_mode_allow {
+            info!(target: "ext_proc", "{} - continue (failure_mode_allow is true)", log_msg);
+            self.request_processing.inject_inflight_frames_and_complete().await;
+            self.request_processing.frame_bridge_close(&mut self.timeout_state.active).await;
+            self.response_processing.inject_inflight_frames_and_complete().await;
+            self.response_processing.frame_bridge_close(&mut self.timeout_state.active).await;
+        } else {
+            info!(target: "ext_proc", "{} - abort (failure_mode_allow is false)", log_msg);
+            _ = self.request_processing.frame_bridge.inject_frame(Err(Box::new(err.clone()))).await;
+            self.request_processing.frame_bridge_close(&mut self.timeout_state.active).await;
+            _ = self.response_processing.frame_bridge.inject_frame(Err(Box::new(err.clone()))).await;
+            self.response_processing.frame_bridge_close(&mut self.timeout_state.active).await;
+        }
+
+        match err {
+            ExtProcError::Timeout(_) => {
+                let status = Action::Return(self.request_processing.status_timeout(self.config.failure_mode_allow));
+                run_action!(self, self.request_processing, status, log_msg);
+                let status = Action::Return(self.response_processing.status_timeout(self.config.failure_mode_allow));
+                run_action!(self, self.response_processing, status, log_msg);
+            },
+            _ => {
+                let status =
+                    Action::Return(self.request_processing.status_error(log_msg, self.config.failure_mode_allow));
+                run_action!(self, self.request_processing, status, log_msg);
+                let status =
+                    Action::Return(self.response_processing.status_error(log_msg, self.config.failure_mode_allow));
+                run_action!(self, self.response_processing, status, log_msg);
+            },
         }
     }
 
@@ -725,6 +832,7 @@ impl ExternalProcessingWorker<kind::Processing> {
                     } else {
                         Either::Right(std::future::pending::<Result<Option<ProcessingResponse>, Status>>())
                     } => {
+
                     debug!(target: "ext_proc", "<= inbound processing response: {inbound_processing_response:?}");
 
                     match inbound_processing_response {
@@ -738,7 +846,7 @@ impl ExternalProcessingWorker<kind::Processing> {
                         Ok(Some(ProcessingResponse { response: Some(ProcessingResponseType::ImmediateResponse(response_attempt)), ..})) => {
                             debug!(target: "ext_proc", "<- ImmediateResponse received");
                             if self.config.disable_immediate_response {
-                                warn!(target: "ext_proc", "External processor attempted to send immediate response which is disabled by config");
+                                info!(target: "ext_proc", "External processor attempted to send immediate response which is disabled by config");
 
                                  if self.config.failure_mode_allow {
                                      with_current_processing!(self, processing,
@@ -803,7 +911,6 @@ impl ExternalProcessingWorker<kind::Processing> {
                                 .handle_body_response(body_response, Some(&self.config.route_cache_action), &mut self.timeout_state.active).await;
 
                             run_action!(self, self.request_processing, action, "body_response");
-
                         },
                         Ok(Some(ProcessingResponse { response: Some(ProcessingResponseType::RequestTrailers(trailers_response)), ..})) => {
                             debug!(target: "ext_proc", "<- RequestTrailers response received");
@@ -844,51 +951,46 @@ impl ExternalProcessingWorker<kind::Processing> {
                             run_action!(self, self.response_processing, action, "handle_trailers_response");
                         },
                         Ok(Some(r)) => {
-                            debug!(target: "ext_proc", "<- unsupported response received: {r:?}");
-
-                            let status = Action::Return(self.request_processing.status_internal_error("unsupported response type received from external processor"));
-                            run_action!(self, self.request_processing, status, "unsupported_response");
-                            let status = Action::Return(self.response_processing.status_internal_error("unsupported response type received from external processor"));
-                            run_action!(self, self.response_processing, status, "unsupported_response");
-
+                            let msg = format!("unsupported response message received: {r:?}");
+                            self.recover_or_failure(ExtProcError::UnsupportedResponseType(format!("{r:?}")), &msg).await;
                             break 'transaction_loop;
                         }
-                        Err(e) => {
-                            let msg = format!("gRPC error received from external processor: {}", e.message());
-                            warn!(target: "ext_proc", msg);
-
-                            let status = Action::Return(self.request_processing.status_error(&msg, self.config.failure_mode_allow));
-                            run_action!(self, self.request_processing, status, "grpc_error");
-                            let status = Action::Return(self.response_processing.status_error(&msg, self.config.failure_mode_allow));
-                            run_action!(self, self.response_processing, status, "grpc_error");
-
+                        Ok(None) => {
+                            let msg = "stream closed by the external processor";
+                            self.recover_or_failure(ExtProcError::UnexpectedEof, msg).await;
                             break 'transaction_loop;
                         },
-                        _ => {
-                            debug!(target: "ext_proc", "stream closed by the external processor");
+                        Err(e) => {
+                            let msg = format!("gRPC error received from external processor: {}", e.message());
+                            self.recover_or_failure(ExtProcError::GrpcError(e.message().into()), &msg).await;
                             break 'transaction_loop;
                         }
                     }
                 },
 
-                outbound_request_body_frame =
-                    async { self.request_processing.frame_bridge.next().await }, if outbound_req_enabled => {
-
+                outbound_request_body_frame = self.request_processing.frame_bridge.next(), if outbound_req_enabled => {
                     debug!(target: "ext_proc", "outbound request body frame: {outbound_request_body_frame:?}");
                     match outbound_request_body_frame {
-                        Some(Ok(current_frame)) => {
-                            // 1. send previously parked frame...
-                            //
-                            if let Some(prev_frame) = self.request_processing.parked_frame.take() {
-                                debug!(target: "ext_proc", "sending body chunk of request ({})",  if prev_frame.is_data() { "DATA" } else { "TRAILERS" });
-                                let action = self.request_processing.handle_outgoing_body_chunk(clone_frame(&prev_frame), false);
-                                run_action!(self, self.request_processing, action, "handle_body_chunk (parked)");
-                                // save a copy of the frame to inject into the body bridge later
-                                self.request_processing.inflight_frames.push(prev_frame);
+                        Some(Ok(frame)) => {
+                            let body_mode = self.overridable_modes.request.body_mode();
+
+                            debug!(target: "ext_proc", "outbound request body frame: buffering frame...");
+                            if let Some(frame_to_send) = self.request_processing.frames_buffer.park(frame) {
+                                // invariant: frame_to_send is always a DATA frame at this point. TRAILERS are sent later.
+                                match body_mode {
+                                    OverridableBodyMode::None => { // body processing is disabled, just inject back the frame
+                                        debug!(target: "ext_proc", "outbound request body frame: injecting the frame DATA into the body");
+                                        _ = self.request_processing.frame_bridge.inject_frame(Ok(frame_to_send)).await;
+                                    },
+                                    _ => { // send the merged frame to ext_proc and park a copy for later injection
+                                        debug!(target: "ext_proc", "outbound request body frame: sending body chunk of request ({})",  if frame_to_send.is_data() { "DATA" } else { "TRAILERS" });
+                                        let action = self.request_processing.handle_outgoing_body_chunk(clone_frame(&frame_to_send), false);
+                                        run_action!(self, self.request_processing, action, "handle_body_chunk (merged frame sent)");
+                                        // save a copy of the frame to inject into the body bridge later
+                                        self.request_processing.inflight_frames.push(frame_to_send);
+                                    }
+                                }
                             }
-                            // 2. park this frame for delayed transmission or inject it directly into the body bridge...
-                            //
-                            self.request_processing.park_or_inject_frame(current_frame, &self.overridable_modes).await;
                         },
                         Some(Err(_err)) => {
                             request_body_to_ext_proc_complete = true;
@@ -897,45 +999,68 @@ impl ExternalProcessingWorker<kind::Processing> {
                         },
                         None => {
                             request_body_to_ext_proc_complete = true;
-                            debug!(target: "ext_proc", "request body stream ended!");
-                            if let Some(current_frame) = self.request_processing.parked_frame.take() {
-                                debug!(target: "ext_proc", "sending last body chunk of request!");
-                                let action = self.request_processing.handle_outgoing_body_chunk(clone_frame(&current_frame), true);
-                                run_action!(self, self.request_processing, action, "handle_body_chunk (end)");
-                                self.request_processing.inflight_frames.push(current_frame);
+                            debug!(target: "ext_proc", "request body stream completing...");
+
+                            while let Some(last_frame) = self.request_processing.frames_buffer.take() {
+                                if last_frame.is_data() { // DATA
+                                  if self.overridable_modes.request.should_process_body() {
+                                      debug!(target: "ext_proc", "sending the last body chunk of request");
+                                      let end_of_stream = !self.request_processing.frames_buffer.has_trailers();
+                                      let action = self.request_processing.handle_outgoing_body_chunk(clone_frame(&last_frame), end_of_stream);
+                                      run_action!(self, self.request_processing, action, "handle_body_chunk (last frame sent)");
+                                      // save a copy of the frame to inject into the body bridge later
+                                      self.request_processing.inflight_frames.push(last_frame);
+                                  } else {
+                                      debug!(target: "ext_proc", "injecting the last body chunk of request into the body");
+                                      _ = self.request_processing.frame_bridge.inject_frame(Ok(last_frame)).await;
+                                  }
+                                } else { // TRAILERS
+                                    if self.overridable_modes.request.should_process_trailers() {
+                                        debug!(target: "ext_proc", "sending the last body chunk of request");
+                                        let action = self.request_processing.handle_outgoing_body_chunk(clone_frame(&last_frame), true);
+                                        run_action!(self, self.request_processing, action, "handle_body_chunk (last frame sent)");
+                                        // save a copy of the frame to inject into the body bridge later
+                                        self.request_processing.inflight_frames.push(last_frame);
+                                    } else {
+                                        debug!(target: "ext_proc", "injecting the last body chunk of request into the body");
+                                        self.request_processing.parked_trailers = Some(last_frame);
+                                    }
+                                }
                             }
 
                             self.request_processing.end_of_stream = true;
 
                             if self.request_processing.inflight_frames.is_empty() {
                                 debug!(target: "ext_proc", "frame bridge closed (request body)!");
-                                self.request_processing.frame_bridge_close(&mut self.timeout_state.active);
+                                self.request_processing.frame_bridge_close(&mut self.timeout_state.active).await;
                             }
                         }
                     }
                 },
 
-                outbound_response_body_frame =
-                    async { self.response_processing.frame_bridge.next().await }, if outbound_resp_enabled => {
-
+                outbound_response_body_frame = self.response_processing.frame_bridge.next(), if outbound_resp_enabled => {
                     debug!(target: "ext_proc", "outbound response body frame: {outbound_response_body_frame:?}");
                     match outbound_response_body_frame {
-                        Some(Ok(current_frame)) => {
+                        Some(Ok(frame)) => {
+                            let body_mode = self.overridable_modes.response.body_mode();
 
-                            // 1. send previously parked frame...
-                            //
-                            if let Some(prev_frame) = self.response_processing.parked_frame.take() {
-                                debug!(target: "ext_proc", "sending body chunk of response ({})",  if prev_frame.is_data() { "DATA" } else { "TRAILERS" });
-                                let action = self.response_processing.handle_outgoing_body_chunk(clone_frame(&prev_frame), false);
-                                run_action!(self, self.response_processing, action, "handle_body_chunk (parked)");
-                                // save a copy of the frame to inject into the body bridge later
-                                self.response_processing.inflight_frames.push(prev_frame);
+                            debug!(target: "ext_proc", "outbound response body frame: buffering frame...");
+                            if let Some(frame_to_send) = self.response_processing.frames_buffer.park(frame) {
+                                // invariant: frame_to_send is always a DATA frame at this point. TRAILERS are sent later.
+                                match body_mode {
+                                    OverridableBodyMode::None => { // body processing is disabled, just inject back the frame
+                                        debug!(target: "ext_proc", "outbound response body frame: injecting the frame DATA into the body");
+                                        _ = self.response_processing.frame_bridge.inject_frame(Ok(frame_to_send)).await;
+                                    },
+                                    _ => { // send the merged frame to ext_proc and park a copy for later injection
+                                        debug!(target: "ext_proc", "outbound response body frame: sending body chunk of response ({})",  if frame_to_send.is_data() { "DATA" } else { "TRAILERS" });
+                                        let action = self.response_processing.handle_outgoing_body_chunk(clone_frame(&frame_to_send), false);
+                                        run_action!(self, self.response_processing, action, "handle_body_chunk (merged frame sent)");
+                                        // save a copy of the frame to inject into the body bridge later
+                                        self.response_processing.inflight_frames.push(frame_to_send);
+                                    }
+                                }
                             }
-
-                            // 2. park this frame for delayed transmission or inject it directly into the body bridge...
-                            //
-                            self.response_processing.park_or_inject_frame(current_frame, &self.overridable_modes).await;
-
                         },
                         Some(Err(_err)) => {
                             response_body_to_ext_proc_complete = true;
@@ -944,48 +1069,49 @@ impl ExternalProcessingWorker<kind::Processing> {
                         },
                         None => {
                             response_body_to_ext_proc_complete = true;
-                            debug!(target: "ext_proc", "response body stream ended!");
-                            if let Some(current_frame) = self.response_processing.parked_frame.take() {
-                                debug!(target: "ext_proc", "sending last body chunk of response!");
-                                let action = self.response_processing.handle_outgoing_body_chunk(clone_frame(&current_frame), true);
-                                run_action!(self, self.response_processing, action, "handle_body_chunk (end)");
-                                self.response_processing.inflight_frames.push(current_frame);
+                            debug!(target: "ext_proc", "response body stream completing...");
+
+                            while let Some(last_frame) = self.response_processing.frames_buffer.take() {
+                                if last_frame.is_data() { // DATA
+                                  if self.overridable_modes.response.should_process_body() {
+                                      debug!(target: "ext_proc", "sending the last body chunk of response");
+                                      let end_of_stream = !self.response_processing.frames_buffer.has_trailers();
+                                      let action = self.response_processing.handle_outgoing_body_chunk(clone_frame(&last_frame), end_of_stream);
+                                      run_action!(self, self.response_processing, action, "handle_body_chunk (last frame sent)");
+                                      // save a copy of the frame to inject into the body bridge later
+                                      self.response_processing.inflight_frames.push(last_frame);
+                                  } else {
+                                      debug!(target: "ext_proc", "injecting the last body chunk of response into the body");
+                                      _ = self.response_processing.frame_bridge.inject_frame(Ok(last_frame)).await;
+                                  }
+                                } else { // TRAILERS
+                                    if self.overridable_modes.response.should_process_trailers() {
+                                        debug!(target: "ext_proc", "sending the last body chunk of response");
+                                        let action = self.response_processing.handle_outgoing_body_chunk(clone_frame(&last_frame), true);
+                                        run_action!(self, self.response_processing, action, "handle_body_chunk (last frame sent)");
+                                        // save a copy of the frame to inject into the body bridge later
+                                        self.response_processing.inflight_frames.push(last_frame);
+                                    } else {
+                                        debug!(target: "ext_proc", "injecting the last body chunk of response into the body");
+                                        self.response_processing.parked_trailers = Some(last_frame);
+                                    }
+                                }
                             }
 
                             self.response_processing.end_of_stream = true;
 
                             if self.response_processing.inflight_frames.is_empty() {
                                 debug!(target: "ext_proc", "frame bridge closed (response body)!");
-                                self.response_processing.frame_bridge_close(&mut self.timeout_state.active);
-                            }
+                                self.response_processing.frame_bridge_close(&mut self.timeout_state.active).await;
+                           }
                         }
                     }
                 },
 
                 () = fast_timeout::fast_sleep(self.timeout_state.duration), if self.timeout_state.active => {
-                    debug!(target: "ext_proc", "loop: message timeout!");
-
-                    if self.config.failure_mode_allow {
-                        warn!(target: "ext_proc", "external processing message timeout - continue (failure_mode_allow is true)");
-
-                        self.request_processing.inject_inflight_frames_and_complete().await;
-                        self.request_processing.frame_bridge_close(&mut self.timeout_state.active);
-                        self.response_processing.inject_inflight_frames_and_complete().await;
-                        self.response_processing.frame_bridge_close(&mut self.timeout_state.active);
-
-                    } else {
-                        warn!(target: "ext_proc", "external processing message timeout - abort (failure_mode_allow is false)");
-
-                        _ = self.request_processing.frame_bridge.inject_frame(Err(Box::new(ExtProcError::Timeout("request timeout")))).await;
-                        self.request_processing.frame_bridge_close(&mut self.timeout_state.active);
-                        _ = self.response_processing.frame_bridge.inject_frame(Err(Box::new(ExtProcError::Timeout("request timeout")))).await;
-                        self.response_processing.frame_bridge_close(&mut self.timeout_state.active);
-                    }
-
-                    let status = Action::Return(self.request_processing.status_timeout(self.config.failure_mode_allow));
-                    run_action!(self, self.request_processing, status, "message_timeout");
-                    let status = Action::Return(self.response_processing.status_timeout(self.config.failure_mode_allow));
-                    run_action!(self, self.response_processing, status, "message_timeout");
+                    let msg = "processing_loop: message timeout";
+                    self.recover_or_failure(ExtProcError::Timeout("message timeout"), &msg).await;
+                    break 'transaction_loop;
                 }
             }
         }
@@ -1054,19 +1180,23 @@ impl ExternalProcessingWorker<kind::Observability> {
                     }
                 },
 
-                outbound_request_body_frame =
-                    async { self.request_processing.frame_bridge.next().await }, if outbound_req_enabled => {
-
+                outbound_request_body_frame = self.request_processing.frame_bridge.next(), if outbound_req_enabled => {
                     debug!(target: "ext_proc", "outbound request body frame: {outbound_request_body_frame:?}");
                     match outbound_request_body_frame {
                         Some(Ok(frame)) => {
-                            // send the current frame and inject back into the body bridge...
-                            debug!(target: "ext_proc", "sending body chunk of request ({})",  if frame.is_data() { "DATA" } else { "TRAILERS" });
-                            let action = self.request_processing.handle_outgoing_body_chunk(clone_frame(&frame), false);
-                            run_action!(self, self.request_processing, action, "handle_body_chunk");
-                            _ = self.request_processing.frame_bridge.inject_frame(Ok(frame)).await;
-                            let action = Action::Return(ProcessingStatus::RequestReady(ReadyStatus::default()));
-                            run_action!(self, self.request_processing, action, "handle_body_chunk (returning after first chunk)");
+                            debug!(target: "ext_proc", "outbound request body frame: buffering frame...");
+                            if let Some(frame_to_send) = self.request_processing.frames_buffer.park(frame) {
+                                if self.overridable_modes.request.should_process_body() {
+                                    _ = self.request_processing.frame_bridge.inject_frame(Ok(clone_frame(&frame_to_send))).await;
+                                    debug!(target: "ext_proc", "outbound request body frame: sending body chunk of request ({})",  if frame_to_send.is_data() { "DATA" } else { "TRAILERS" });
+                                    let action = self.request_processing.handle_outgoing_body_chunk(frame_to_send, false);
+                                    run_action!(self, self.request_processing, action, "handle_body_chunk (merged frame sent)");
+
+                                } else {
+                                    debug!(target: "ext_proc", "outbound request body frame: injecting the frame DATA into the body");
+                                    _ = self.request_processing.frame_bridge.inject_frame(Ok(frame_to_send)).await;
+                                }
+                            }
                         },
                         Some(Err(_err)) => {
                             request_body_to_ext_proc_complete = true;
@@ -1074,29 +1204,58 @@ impl ExternalProcessingWorker<kind::Observability> {
                             self.request_processing.status_error("error occurred when streaming request body to external processing", self.config.failure_mode_allow);
                         },
                         None => {
-                            debug!(target: "ext_proc", "request body stream ended!");
                             request_body_to_ext_proc_complete = true;
-                            self.request_processing.end_of_stream = true;
-                            self.request_processing.frame_bridge_close(&mut self.timeout_state.active);
+                            debug!(target: "ext_proc", "request body stream completing...");
+
+                            while let Some(last_frame) = self.request_processing.frames_buffer.take() {
+                                if last_frame.is_data() { // DATA
+                                  if self.overridable_modes.request.should_process_body() {
+                                      debug!(target: "ext_proc", "sending the last data chunk of request");
+                                      _ = self.request_processing.frame_bridge.inject_frame(Ok(clone_frame(&last_frame))).await;
+                                      let end_of_stream = !self.request_processing.frames_buffer.has_trailers();
+                                      let action = self.request_processing.handle_outgoing_body_chunk(last_frame, end_of_stream);
+                                      run_action!(self, self.request_processing, action, "handle_body_chunk (last data frame sent)");
+                                  } else {
+                                      debug!(target: "ext_proc", "injecting the last data chunk of request into the body");
+                                      _ = self.request_processing.frame_bridge.inject_frame(Ok(last_frame)).await;
+                                  }
+                                } else { // TRAILERS
+                                    if self.overridable_modes.request.should_process_trailers() {
+                                        debug!(target: "ext_proc", "sending trailers chunk of request");
+                                        _ = self.request_processing.frame_bridge.inject_frame(Ok(clone_frame(&last_frame))).await;
+                                        let action = self.request_processing.handle_outgoing_body_chunk(last_frame, true);
+                                        run_action!(self, self.request_processing, action, "handle_body_chunk (trailer frame sent)");
+                                    } else {
+                                        debug!(target: "ext_proc", "injecting the trailers chunk of request into the body");
+                                        _ = self.request_processing.frame_bridge.inject_frame(Ok(last_frame)).await;
+                                    }
+                                }
+                            }
+
                             let action = Action::Return(ProcessingStatus::RequestReady(ReadyStatus::default()));
-                            run_action!(self, self.request_processing, action, "handle_body_chunk (end)");
+                            run_action!(self, self.request_processing, action, "request streaming completed");
+                            self.request_processing.end_of_stream = true;
                         }
                     }
                 },
 
-                outbound_response_body_frame =
-                    async { self.response_processing.frame_bridge.next().await }, if outbound_resp_enabled => {
-
+                outbound_response_body_frame = self.response_processing.frame_bridge.next(), if outbound_resp_enabled => {
                     debug!(target: "ext_proc", "outbound response body frame: {outbound_response_body_frame:?}");
                     match outbound_response_body_frame {
                         Some(Ok(frame)) => {
-                            // send the current frame and inject back into the body bridge...
-                            debug!(target: "ext_proc", "sending body chunk of request ({})",  if frame.is_data() { "DATA" } else { "TRAILERS" });
-                            let action = self.response_processing.handle_outgoing_body_chunk(clone_frame(&frame), false);
-                            run_action!(self, self.response_processing, action, "handle_body_chunk");
-                            _ = self.response_processing.frame_bridge.inject_frame(Ok(frame)).await;
-                            let action = Action::Return(ProcessingStatus::ResponseReady(ReadyStatus::default()));
-                            run_action!(self, self.response_processing, action, "handle_body_chunk (returning after first chunk)");
+                            debug!(target: "ext_proc", "outbound response body frame: buffering frame...");
+                            if let Some(frame_to_send) = self.response_processing.frames_buffer.park(frame) {
+                                if self.overridable_modes.response.should_process_body() {
+                                    _ = self.response_processing.frame_bridge.inject_frame(Ok(clone_frame(&frame_to_send))).await;
+                                    debug!(target: "ext_proc", "outbound response body frame: sending body chunk of response ({})",  if frame_to_send.is_data() { "DATA" } else { "TRAILERS" });
+                                    let action = self.response_processing.handle_outgoing_body_chunk(frame_to_send, false);
+                                    run_action!(self, self.response_processing, action, "handle_body_chunk (merged frame sent)");
+
+                                } else {
+                                    debug!(target: "ext_proc", "outbound response body frame: injecting the frame DATA into the body");
+                                    _ = self.response_processing.frame_bridge.inject_frame(Ok(frame_to_send)).await;
+                                }
+                            }
                         },
                         Some(Err(_err)) => {
                             response_body_to_ext_proc_complete = true;
@@ -1104,14 +1263,44 @@ impl ExternalProcessingWorker<kind::Observability> {
                             self.response_processing.status_error("error occurred when streaming response body to external processing", self.config.failure_mode_allow);
                         },
                         None => {
-                            debug!(target: "ext_proc", "request body stream ended!");
                             response_body_to_ext_proc_complete = true;
-                            self.response_processing.end_of_stream = true;
-                            self.response_processing.frame_bridge_close(&mut self.timeout_state.active);
+                            debug!(target: "ext_proc", "response body stream completing...");
+
+                            while let Some(last_frame) = self.response_processing.frames_buffer.take() {
+                                if last_frame.is_data() { // DATA
+                                  if self.overridable_modes.response.should_process_body() {
+                                      debug!(target: "ext_proc", "sending the last data chunk of response");
+                                      _ = self.response_processing.frame_bridge.inject_frame(Ok(clone_frame(&last_frame))).await;
+                                      let end_of_stream = !self.response_processing.frames_buffer.has_trailers();
+                                      let action = self.response_processing.handle_outgoing_body_chunk(last_frame, end_of_stream);
+                                      run_action!(self, self.response_processing, action, "handle_body_chunk (last data frame sent)");
+                                  } else {
+                                      debug!(target: "ext_proc", "injecting the last data chunk of response into the body");
+                                      _ = self.response_processing.frame_bridge.inject_frame(Ok(last_frame)).await;
+                                  }
+                                } else { // TRAILERS
+                                    if self.overridable_modes.response.should_process_trailers() {
+                                        debug!(target: "ext_proc", "sending trailers chunk of response");
+                                        _ = self.response_processing.frame_bridge.inject_frame(Ok(clone_frame(&last_frame))).await;
+                                        let action = self.response_processing.handle_outgoing_body_chunk(last_frame, true);
+                                        run_action!(self, self.response_processing, action, "handle_body_chunk (trailer frame sent)");
+                                    } else {
+                                        debug!(target: "ext_proc", "injecting the trailers chunk of response into the body");
+                                        _ = self.response_processing.frame_bridge.inject_frame(Ok(last_frame)).await;
+                                    }
+                                }
+                            }
+
                             let action = Action::Return(ProcessingStatus::ResponseReady(ReadyStatus::default()));
-                            run_action!(self, self.response_processing, action, "handle_body_chunk (end)");
+                            run_action!(self, self.response_processing, action, "response streaming completed");
+                            self.response_processing.end_of_stream = true;
                         }
                     }
+                },
+
+                _ = std::future::ready(()), if streaming_enabled && !outbound_req_enabled && !outbound_resp_enabled => {
+                    debug!(target: "ext_proc", "no outbound request or response enabled (observability terminating...)");
+                    break 'transaction_loop;
                 }
             }
         }
@@ -1194,7 +1383,7 @@ impl<S: kind::Mode + Default> ExternalProcessingWorker<S> {
 
         match send_outcome {
             Some(Err(err)) => {
-                warn!(target: "ext_proc", "External processor is unavailable: {err}");
+                info!(target: "ext_proc", "External processor is unavailable: {err}");
                 if let Some(reply_channel) = self.response_processing.reply_channel.take() {
                     let _ = reply_channel.send(
                         self.response_processing
@@ -1235,12 +1424,12 @@ impl<S: kind::Mode + Default> ExternalProcessingWorker<S> {
         let mut timeout_duration =
             Duration::from_secs(extended_timeout.seconds as u64) + Duration::from_nanos(extended_timeout.nanos as u64);
         if timeout_duration < Duration::from_millis(1) {
-            warn!(target:"ext_proc", "External processor: override_message_timeout must be >= 1ms");
+            info!(target:"ext_proc", "External processor: override_message_timeout must be >= 1ms");
             timeout_duration = self.config.message_timeout;
         }
         if let Some(max_timeout) = self.config.max_message_timeout {
             if timeout_duration > max_timeout {
-                warn!(target:"ext_proc", "External processor: attempted to override message timeout to value > max_message_timeout (defaulting to max_message_timeout)");
+                info!(target:"ext_proc", "External processor: attempted to override message timeout to value > max_message_timeout (defaulting to max_message_timeout)");
                 timeout_duration = max_timeout;
             }
         }
@@ -1258,7 +1447,7 @@ impl<S: kind::Mode + Default> ExternalProcessingWorker<S> {
             .unwrap_or(http::StatusCode::OK);
         let body_bytes = Bytes::copy_from_slice(&response_attempt.body);
         let body = Full::new(body_bytes);
-        let mut response = Response::new(crate::PolyBody::from(body));
+        let mut response = Response::new(PolyBody::from(body));
         *response.status_mut() = status;
         if let Some(header_mutation) = &response_attempt.headers {
             let _ =
@@ -1270,2114 +1459,5 @@ impl<S: kind::Mode + Default> ExternalProcessingWorker<S> {
             }
         }
         response
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{
-        body::{body_with_metrics::BodyWithMetrics, response_flags::BodyKind},
-        listeners::http_connection_manager::ext_proc::{
-            kind::{MsgKind, RequestMsg, ResponseMsg},
-            mutation::apply_header_mutations,
-            r#override::ModeSelector,
-        },
-    };
-    use http::{Method, StatusCode, Version};
-    use http_body_util::{BodyExt, Empty, Full};
-    use orion_configuration::config::network_filters::http_connection_manager::http_filters::ext_proc::{
-        BodyProcessingMode, ExternalProcessor as ExternalProcessorConfig, GoogleGrpc, GrpcService,
-        GrpcServiceSpecifier, HeaderProcessingMode, ProcessingMode, RouteCacheAction, TrailerProcessingMode,
-    };
-    use orion_data_plane_api::envoy_data_plane_api::envoy::extensions::filters::http::ext_proc::v3::{
-        processing_mode, ProcessingMode as EnvoyProcessingMode,
-    };
-    use orion_data_plane_api::envoy_data_plane_api::{
-        envoy::{
-            config::core::v3::{
-                header_value_option::HeaderAppendAction, HeaderValue as EnvoyHeaderValue, HeaderValueOption,
-            },
-            r#type::v3::HttpStatus as EnvoyHttpStatus,
-            service::ext_proc::v3::{
-                body_mutation::Mutation,
-                common_response::ResponseStatus,
-                external_processor_server::{ExternalProcessor as ExternalProcessorService, ExternalProcessorServer},
-                processing_response::Response as ProcessingResponseType,
-                BodyMutation, BodyResponse, CommonResponse, HeaderMutation, HeadersResponse, ProcessingRequest,
-                ProcessingResponse, StreamedBodyResponse, TrailersResponse,
-            },
-        },
-        tonic::{
-            async_trait,
-            transport::{Error as TonicError, Server},
-            Request as TonicRequest, Response as TonicResponse, Status, Streaming,
-        },
-    };
-    use std::{collections::VecDeque, future::ready, net::SocketAddr, str::FromStr, time::Duration};
-    use tokio::{net::TcpListener, task::JoinHandle};
-
-    use tokio_stream::wrappers::{ReceiverStream, TcpListenerStream};
-
-    #[derive(Debug, Clone)]
-    pub struct MockExternalProcessorState {
-        responses: VecDeque<ProcessingResponse>,
-    }
-
-    impl MockExternalProcessorState {
-        pub fn new() -> Self {
-            Self { responses: VecDeque::new() }
-        }
-        pub fn add_response(mut self, response: ProcessingResponse) -> Self {
-            self.responses.push_back(response);
-            self
-        }
-        pub fn get_next_response(&mut self) -> Option<ProcessingResponse> {
-            self.responses.pop_front()
-        }
-    }
-
-    #[derive(Debug)]
-    pub struct MockExternalProcessor {
-        state: MockExternalProcessorState,
-        delay: Option<Duration>,
-    }
-
-    impl MockExternalProcessor {
-        pub fn new(state: MockExternalProcessorState, delay: Option<Duration>) -> Self {
-            Self { state, delay }
-        }
-    }
-
-    #[async_trait]
-    impl ExternalProcessorService for MockExternalProcessor {
-        type ProcessStream = ReceiverStream<Result<ProcessingResponse, Status>>;
-
-        async fn process(
-            &self,
-            request: TonicRequest<Streaming<ProcessingRequest>>,
-        ) -> Result<TonicResponse<Self::ProcessStream>, Status> {
-            let mut inbound = request.into_inner();
-            let mut state = self.state.clone();
-            let delay = self.delay.clone();
-            let (tx, rx) = tokio::sync::mpsc::channel(16);
-            tokio::spawn(async move {
-                while let Some(_req) = inbound.message().await.unwrap_or(None) {
-                    if let Some(response) = state.get_next_response() {
-                        if let Some(delay) = delay {
-                            tokio::time::sleep(delay).await;
-                        }
-                        if tx.send(Ok(response)).await.is_err() {
-                            break;
-                        }
-                    }
-                }
-            });
-            let output_stream = ReceiverStream::new(rx);
-            Ok(TonicResponse::new(output_stream))
-        }
-    }
-
-    async fn start_mock_server(state: MockExternalProcessorState) -> (SocketAddr, JoinHandle<Result<(), TonicError>>) {
-        let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let socket_addr = listener.local_addr().unwrap();
-        let incoming = TcpListenerStream::new(listener);
-        let mock_service = MockExternalProcessor::new(state, None);
-        let server =
-            Server::builder().add_service(ExternalProcessorServer::new(mock_service)).serve_with_incoming(incoming);
-        let server_handle = tokio::spawn(server);
-        (socket_addr, server_handle)
-    }
-
-    async fn start_mock_server_with_delay(
-        state: MockExternalProcessorState,
-        delay: Duration,
-    ) -> (SocketAddr, JoinHandle<Result<(), TonicError>>) {
-        let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let socket_addr = listener.local_addr().unwrap();
-        let incoming = TcpListenerStream::new(listener);
-        let mock_service = MockExternalProcessor::new(state, Some(delay));
-        let server =
-            Server::builder().add_service(ExternalProcessorServer::new(mock_service)).serve_with_incoming(incoming);
-        let server_handle = tokio::spawn(server);
-        (socket_addr, server_handle)
-    }
-
-    #[derive(Debug, Default)]
-    struct Mock<M: MsgKind> {
-        headers: Vec<Option<(&'static str, &'static str)>>,
-        body: Option<&'static str>,
-        trailers: Vec<Option<(&'static str, &'static str)>>,
-        _marker: std::marker::PhantomData<M>,
-    }
-
-    impl<M: MsgKind> Mock<M> {
-        fn new(
-            headers: Vec<Option<(&'static str, &'static str)>>,
-            body: Option<&'static str>,
-            trailers: Vec<Option<(&'static str, &'static str)>>,
-        ) -> Self {
-            Self { headers, body, trailers, _marker: std::marker::PhantomData }
-        }
-
-        fn orig_headers_map(&self) -> http::HeaderMap {
-            self.apply_header_mutation(None)
-        }
-
-        fn orig_trailers_map(&self) -> http::HeaderMap {
-            self.apply_trailer_mutation(None)
-        }
-
-        fn apply_header_mutation(&self, header_mutation: Option<&HeaderMutation>) -> http::HeaderMap {
-            let mut header_map = http::HeaderMap::with_capacity(self.headers.len());
-            for (key, value) in self.headers.iter().flatten() {
-                header_map.insert(http::HeaderName::from_static(key), http::HeaderValue::from_static(value));
-            }
-            if let Some(mutation) = header_mutation.as_ref() {
-                apply_header_mutations(&mut header_map, mutation, None).unwrap();
-            }
-            header_map
-        }
-
-        fn apply_body_mutation(&self, body_mutation: Option<&Mutation>) -> Option<Bytes> {
-            let body_replacement: Option<Bytes> = match body_mutation {
-                Some(Mutation::Body(bytes)) => Some(bytes.clone().into()),
-                Some(Mutation::ClearBody(true)) => Some(Bytes::new()),
-                Some(Mutation::ClearBody(false)) | None => None,
-                Some(Mutation::StreamedResponse(chunk)) => Some(chunk.clone().body.into()),
-            };
-            body_replacement.or(self.body.map(Into::into))
-        }
-
-        fn apply_trailer_mutation(&self, trailer_mutation: Option<&HeaderMutation>) -> http::HeaderMap {
-            let mut trailer_map = http::HeaderMap::with_capacity(self.trailers.len());
-            for (key, value) in self.trailers.iter().flatten() {
-                trailer_map.insert(http::HeaderName::from_static(key), http::HeaderValue::from_static(value));
-            }
-            if let Some(mutation) = trailer_mutation.as_ref() {
-                apply_header_mutations(&mut trailer_map, mutation, None).unwrap();
-            }
-            trailer_map
-        }
-    }
-
-    fn build_request_from_mock(mock_request: &Mock<RequestMsg>) -> Request<BodyWithMetrics<PolyBody>> {
-        let mut req = Request::builder().method(Method::GET).uri("http://example.com/test").version(Version::HTTP_11);
-
-        if let Some(headers) = transform(mock_request.headers.clone()) {
-            for (name, value) in headers {
-                req = req.header(name, value);
-            }
-        }
-
-        let mut trailers_map = http::HeaderMap::with_capacity(mock_request.trailers.len());
-        if let Some(trailers) = transform(mock_request.trailers.clone()) {
-            trailers_map = if trailers.is_empty() {
-                http::HeaderMap::default()
-            } else {
-                let mut map = http::header::HeaderMap::new();
-                for (name, value) in trailers {
-                    map.append(
-                        http::header::HeaderName::from_str(name).unwrap(),
-                        http::header::HeaderValue::from_str(value).unwrap().clone(),
-                    );
-                }
-                map
-            };
-        }
-
-        let body = match mock_request.body {
-            None if !trailers_map.is_empty() => PolyBody::from(
-                Empty::<bytes::Bytes>::new().with_trailers(ready(Some(Ok::<_, Infallible>(trailers_map)))),
-            ),
-            Some(b) if !trailers_map.is_empty() => {
-                PolyBody::from(Full::new(bytes::Bytes::from(b.to_owned())).with_trailers(ready(Some(Ok::<
-                    _,
-                    Infallible,
-                >(
-                    trailers_map
-                )))))
-            },
-            None => PolyBody::from(Empty::<bytes::Bytes>::new()),
-            Some(b) => PolyBody::from(Full::new(bytes::Bytes::from(b.to_owned()))),
-        };
-
-        req.body(BodyWithMetrics::new(BodyKind::Request, body, |_, _, _| {})).unwrap()
-    }
-
-    fn build_response_from_mock(mock_response: &Mock<ResponseMsg>) -> Response<PolyBody> {
-        let mut resp = Response::builder().version(Version::HTTP_11);
-
-        if let Some(headers) = transform(mock_response.headers.clone()) {
-            for (name, value) in headers {
-                resp = resp.header(name, value);
-            }
-        }
-
-        let mut trailers_map = http::HeaderMap::with_capacity(mock_response.trailers.len());
-        if let Some(trailers) = transform(mock_response.trailers.clone()) {
-            trailers_map = if trailers.is_empty() {
-                http::HeaderMap::default()
-            } else {
-                let mut map = http::header::HeaderMap::new();
-                for (name, value) in trailers {
-                    map.append(
-                        http::header::HeaderName::from_str(name).unwrap(),
-                        http::header::HeaderValue::from_str(value).unwrap().clone(),
-                    );
-                }
-                map
-            };
-        }
-
-        let body = match mock_response.body {
-            None if !trailers_map.is_empty() => PolyBody::from(
-                Empty::<bytes::Bytes>::new().with_trailers(ready(Some(Ok::<_, Infallible>(trailers_map)))),
-            ),
-            Some(b) if !trailers_map.is_empty() => {
-                PolyBody::from(Full::new(bytes::Bytes::from(b.to_owned())).with_trailers(ready(Some(Ok::<
-                    _,
-                    Infallible,
-                >(
-                    trailers_map
-                )))))
-            },
-            None => PolyBody::from(Empty::<bytes::Bytes>::new()),
-            Some(b) => PolyBody::from(Full::new(bytes::Bytes::from(b.to_owned()))),
-        };
-
-        resp.body(body).unwrap()
-    }
-
-    fn create_default_config_for_ext_proc_filter(
-        server_addr: SocketAddr,
-        processing_mode: ProcessingMode,
-    ) -> ExternalProcessorConfig {
-        ExternalProcessorConfig {
-            grpc_service: GrpcService {
-                service_specifier: GrpcServiceSpecifier::GoogleGrpc(GoogleGrpc {
-                    target_uri: format!("http://{server_addr}"),
-                }),
-                timeout: Some(Duration::from_secs(1)),
-            },
-            processing_mode: Some(processing_mode),
-            observability_mode: false,
-            failure_mode_allow: false,
-            disable_immediate_response: false,
-            forward_rules: None,
-            mutation_rules: None,
-            message_timeout: Some(Duration::from_millis(200)),
-            max_message_timeout: None,
-            allowed_override_modes: vec![],
-            allow_mode_override: false,
-            route_cache_action: RouteCacheAction::Default,
-            send_body_without_waiting_for_header_response: false,
-            deferred_close_timeout: None,
-        }
-    }
-
-    #[inline]
-    fn create_header_mutation(headers: Vec<(&str, &str)>) -> Option<HeaderMutation> {
-        if headers.is_empty() {
-            None
-        } else {
-            Some(HeaderMutation {
-                set_headers: headers
-                    .into_iter()
-                    .map(|(key, value)| HeaderValueOption {
-                        header: Some(EnvoyHeaderValue {
-                            key: key.to_owned(),
-                            value: value.to_owned(),
-                            raw_value: vec![],
-                        }),
-                        #[allow(deprecated)]
-                        append: None,
-                        append_action: HeaderAppendAction::OverwriteIfExistsOrAdd as i32,
-                        keep_empty_value: false,
-                    })
-                    .collect(),
-                remove_headers: vec![],
-            })
-        }
-    }
-
-    #[inline]
-    fn create_body_mutation(body: Vec<u8>, end_of_stream: Option<bool>) -> BodyMutation {
-        if let Some(end_of_stream) = end_of_stream {
-            BodyMutation { mutation: Some(Mutation::StreamedResponse(StreamedBodyResponse { body, end_of_stream })) }
-        } else {
-            BodyMutation { mutation: Some(Mutation::Body(body)) }
-        }
-    }
-
-    #[inline]
-    fn create_trailer_mutation(trailers: Vec<(&str, &str)>) -> Option<HeaderMutation> {
-        create_header_mutation(trailers)
-    }
-
-    #[inline]
-    fn convert_trailers_to_envoy_header_map(trailers: Vec<(&str, &str)>) -> HeaderMap {
-        HeaderMap {
-            headers: trailers
-                .into_iter()
-                .map(|(key, value)| EnvoyHeaderValue {
-                    key: key.to_owned(),
-                    value: value.to_owned(),
-                    raw_value: vec![],
-                })
-                .collect(),
-        }
-    }
-
-    fn create_immediate_response(
-        headers: Vec<Option<(&str, &str)>>,
-        body_data: Option<Vec<u8>>,
-        status: i32,
-    ) -> ProcessingResponse {
-        let response_headers = transform(headers).and_then(|hdrs| create_header_mutation(hdrs));
-        let response_body = body_data.unwrap_or_default();
-
-        let immediate_response = ImmediateResponse {
-            status: Some(EnvoyHttpStatus { code: status }),
-            headers: response_headers,
-            body: response_body,
-            grpc_status: None,
-            details: "immediate_response".to_string(),
-        };
-
-        ProcessingResponse {
-            response: Some(ProcessingResponseType::ImmediateResponse(immediate_response)),
-            mode_override: None,
-            dynamic_metadata: None,
-            override_message_timeout: None,
-        }
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn create_headers_response<M: MsgKind>(
-        headers: Vec<Option<(&str, &str)>>,
-        body_data: Option<Vec<u8>>,
-        trailers: Vec<Option<(&str, &str)>>,
-        status: i32,
-        end_of_stream: Option<bool>,
-    ) -> ProcessingResponse {
-        let header_mutation = transform(headers).and_then(|hdrs| create_header_mutation(hdrs));
-        let body_mutation = body_data.map(|body| create_body_mutation(body, end_of_stream));
-        let mut trailers_new = None;
-        if status == ResponseStatus::ContinueAndReplace as i32 {
-            if let Some(trailers) = transform(trailers) {
-                trailers_new = Some(convert_trailers_to_envoy_header_map(trailers));
-            }
-        }
-
-        let header_response = HeadersResponse {
-            response: Some(CommonResponse {
-                status,
-                header_mutation,
-                body_mutation,
-                trailers: trailers_new,
-                clear_route_cache: false,
-            }),
-        };
-
-        let response = if M::IS_RESPONSE {
-            Some(ProcessingResponseType::ResponseHeaders(header_response))
-        } else {
-            Some(ProcessingResponseType::RequestHeaders(header_response))
-        };
-
-        ProcessingResponse { response, mode_override: None, dynamic_metadata: None, override_message_timeout: None }
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn create_body_response<M: MsgKind>(
-        headers: Vec<Option<(&str, &str)>>,
-        body_data: Option<Vec<u8>>,
-        trailers: Vec<Option<(&str, &str)>>,
-        status: i32,
-        end_of_stream: Option<bool>,
-    ) -> ProcessingResponse {
-        let header_mutation = transform(headers).and_then(|hdrs| create_header_mutation(hdrs));
-        let body_mutation = body_data.map(|body| create_body_mutation(body, end_of_stream));
-        let mut trailers_new = None;
-        if status == ResponseStatus::ContinueAndReplace as i32 {
-            if let Some(trailers) = transform(trailers) {
-                trailers_new = Some(convert_trailers_to_envoy_header_map(trailers));
-            }
-        }
-
-        let body_response = BodyResponse {
-            response: Some(CommonResponse {
-                status,
-                header_mutation,
-                body_mutation,
-                trailers: trailers_new,
-                clear_route_cache: false,
-            }),
-        };
-
-        let response = if M::IS_RESPONSE {
-            Some(ProcessingResponseType::ResponseBody(body_response))
-        } else {
-            Some(ProcessingResponseType::RequestBody(body_response))
-        };
-
-        ProcessingResponse { response, mode_override: None, dynamic_metadata: None, override_message_timeout: None }
-    }
-
-    fn create_trailers_response<M: MsgKind>(trailers: Vec<Option<(&str, &str)>>) -> ProcessingResponse {
-        let trailer_mutation = transform(trailers).and_then(|trls| create_trailer_mutation(trls));
-
-        let trailers_response = TrailersResponse { header_mutation: trailer_mutation };
-        let response = if M::IS_RESPONSE {
-            Some(ProcessingResponseType::ResponseTrailers(trailers_response))
-        } else {
-            Some(ProcessingResponseType::RequestTrailers(trailers_response))
-        };
-        ProcessingResponse { response, mode_override: None, dynamic_metadata: None, override_message_timeout: None }
-    }
-
-    static HEADER_PROCESSING_MODE: [HeaderProcessingMode; 3] =
-        [HeaderProcessingMode::Default, HeaderProcessingMode::Send, HeaderProcessingMode::Skip];
-    static BODY_PROCESSING_MODE: [BodyProcessingMode; 5] = [
-        BodyProcessingMode::None,
-        BodyProcessingMode::Streamed,
-        BodyProcessingMode::Buffered,
-        BodyProcessingMode::BufferedPartial,
-        BodyProcessingMode::FullDuplexStreamed,
-    ];
-    static TRAILER_PROCESSING_MODE: [TrailerProcessingMode; 3] =
-        [TrailerProcessingMode::Default, TrailerProcessingMode::Skip, TrailerProcessingMode::Send];
-
-    fn generate_processing_mode_configurations<M: MsgKind>() -> Vec<ProcessingMode> {
-        if M::IS_RESPONSE {
-            HEADER_PROCESSING_MODE
-                .iter()
-                .flat_map(|&header_mode| {
-                    BODY_PROCESSING_MODE.iter().flat_map(move |&body_mode| {
-                        TRAILER_PROCESSING_MODE.iter().map(move |&trailer_mode| ProcessingMode {
-                            request_header_mode: HeaderProcessingMode::Skip,
-                            request_body_mode: BodyProcessingMode::None,
-                            request_trailer_mode: TrailerProcessingMode::Skip,
-                            response_header_mode: header_mode,
-                            response_body_mode: body_mode,
-                            response_trailer_mode: trailer_mode,
-                        })
-                    })
-                })
-                .collect()
-        } else {
-            HEADER_PROCESSING_MODE
-                .iter()
-                .flat_map(|&header_mode| {
-                    BODY_PROCESSING_MODE.iter().flat_map(move |&body_mode| {
-                        TRAILER_PROCESSING_MODE.iter().map(move |&trailer_mode| ProcessingMode {
-                            request_header_mode: header_mode,
-                            request_body_mode: body_mode,
-                            request_trailer_mode: trailer_mode,
-                            response_header_mode: HeaderProcessingMode::Skip,
-                            response_body_mode: BodyProcessingMode::None,
-                            response_trailer_mode: TrailerProcessingMode::Skip,
-                        })
-                    })
-                })
-                .collect()
-        }
-    }
-
-    // we always include extra headers in the generated request
-    static REQUEST_HEADERS: [Option<(&str, &str)>; 2] = [None, Some(("x-test-header", "original-header-value"))];
-    static REQUEST_BODIES: [Option<&str>; 2] = [None, Some("original body")];
-    static REQUEST_TRAILERS: [Option<(&str, &str)>; 2] = [None, Some(("x-test-trailer", "original-trailer-value"))];
-
-    fn generate_mock_messages<M: MsgKind>() -> Vec<Mock<M>> {
-        REQUEST_HEADERS
-            .iter()
-            .flat_map(|&headers| {
-                REQUEST_BODIES.iter().flat_map(move |&body| {
-                    REQUEST_TRAILERS.iter().map(move |&trailers| Mock::<M> {
-                        headers: vec![headers],
-                        body,
-                        trailers: vec![trailers],
-                        _marker: std::marker::PhantomData,
-                    })
-                })
-            })
-            .collect()
-    }
-
-    static HEADER_MODIFICATIONS: [Option<(&str, &str)>; 2] = [None, Some(("x-test-header", "ext-proc-header-value"))];
-    static BODY_MODIFICATIONS: [Option<&str>; 2] = [None, Some("external processor body")];
-    static TRAILER_MODIFICATIONS: [Option<(&str, &str)>; 2] =
-        [None, Some(("x-test-trailer", "ext-proc-trailer-value"))];
-
-    fn transform<T>(vec: Vec<Option<T>>) -> Option<Vec<T>> {
-        let filtered: Vec<T> = vec.into_iter().flatten().collect();
-        (!filtered.is_empty()).then_some(filtered)
-    }
-
-    fn generate_header_processing_response<M: MsgKind>(status: ResponseStatus) -> Vec<ProcessingResponse> {
-        HEADER_MODIFICATIONS
-            .iter()
-            .flat_map(|&headers| {
-                BODY_MODIFICATIONS.iter().flat_map(move |&body| {
-                    TRAILER_MODIFICATIONS.iter().map(move |&trailers| {
-                        create_headers_response::<M>(
-                            vec![headers],
-                            body.map(Into::into),
-                            // trailers modifications are NYI in body response in Envoy but Orion supports them
-                            vec![trailers],
-                            status as i32,
-                            None,
-                        )
-                    })
-                })
-            })
-            .collect()
-    }
-
-    fn generate_body_processing_response<M: MsgKind>(status: ResponseStatus) -> Vec<ProcessingResponse> {
-        HEADER_MODIFICATIONS
-            .iter()
-            .flat_map(|&headers| {
-                BODY_MODIFICATIONS.iter().flat_map(move |&body| {
-                    TRAILER_MODIFICATIONS.iter().map(move |&trailers| {
-                        create_body_response::<M>(
-                            vec![headers],
-                            body.map(Into::into),
-                            // trailers modifications are NYI in body response in Envoy but Orion supports them
-                            vec![trailers],
-                            status as i32,
-                            None,
-                        )
-                    })
-                })
-            })
-            .collect()
-    }
-
-    fn generate_trailer_processing_response<M: MsgKind>() -> Vec<ProcessingResponse> {
-        TRAILER_MODIFICATIONS.iter().map(move |&trailers| create_trailers_response::<M>(vec![trailers])).collect()
-    }
-
-    fn generate_mock_external_processors_states<M: MsgKind>(status: ResponseStatus) -> Vec<MockExternalProcessorState> {
-        let mut mock_ext_proc_state: Vec<MockExternalProcessorState> = vec![];
-        for header_response in generate_header_processing_response::<M>(status) {
-            for body_response in generate_body_processing_response::<M>(status) {
-                for trailer_response in generate_trailer_processing_response::<M>() {
-                    mock_ext_proc_state.push(
-                        MockExternalProcessorState::new()
-                            .add_response(header_response.clone())
-                            .add_response(body_response.clone())
-                            .add_response(trailer_response),
-                    );
-                }
-            }
-        }
-        mock_ext_proc_state
-    }
-
-    fn validate_mock_server_configuration<M: MsgKind + ModeSelector>(
-        mock_msg: &Mock<M>,
-        processing_mode: &ProcessingMode,
-        mock_ext_proc_state: &MockExternalProcessorState,
-    ) -> bool {
-        let has_headers = transform(mock_msg.headers.clone()).is_some();
-        let has_body = mock_msg.body.is_some();
-        let has_trailers = transform(mock_msg.trailers.clone()).is_some();
-
-        let should_send_headers =
-            has_headers && !matches!(<M as ModeSelector>::header_mode(processing_mode), HeaderProcessingMode::Skip);
-        let should_send_body =
-            has_body && !matches!(<M as ModeSelector>::body_mode(processing_mode), BodyProcessingMode::None);
-        let should_send_trailers =
-            has_trailers && matches!(<M as ModeSelector>::trailer_mode(processing_mode), TrailerProcessingMode::Send);
-
-        for response in &mock_ext_proc_state.responses {
-            match response.response.as_ref().unwrap() {
-                ProcessingResponseType::RequestHeaders(_) | ProcessingResponseType::ResponseHeaders(_) => {
-                    if !should_send_headers {
-                        return false;
-                    }
-                },
-                ProcessingResponseType::RequestBody(_) | ProcessingResponseType::ResponseBody(_) => {
-                    if !should_send_body {
-                        return false;
-                    }
-                },
-                ProcessingResponseType::RequestTrailers(_) | ProcessingResponseType::ResponseTrailers(_) => {
-                    if !should_send_trailers {
-                        return false;
-                    }
-                },
-                ProcessingResponseType::ImmediateResponse(_) => return false,
-            }
-        }
-        true
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn assert_result<M>(
-        test_case_num: i32,
-        mock: &Mock<M>,
-        headers: &http::HeaderMap,
-        body: Option<&Bytes>,
-        trailers: &http::HeaderMap,
-        mock_state: &MockExternalProcessorState,
-        processing_mode: &ProcessingMode,
-    ) where
-        M: std::fmt::Debug + MsgKind + ModeSelector,
-    {
-        // ProcessingMode predicates
-        let has_headers = transform(mock.headers.clone()).is_some();
-        let has_body = mock.body.is_some();
-        let has_trailers = transform(mock.trailers.clone()).is_some();
-
-        let should_send_headers =
-            has_headers && !matches!(<M as ModeSelector>::header_mode(processing_mode), HeaderProcessingMode::Skip);
-        let should_send_body =
-            has_body && !matches!(<M as ModeSelector>::body_mode(processing_mode), BodyProcessingMode::None);
-        let should_send_trailers =
-            has_trailers && matches!(<M as ModeSelector>::trailer_mode(processing_mode), TrailerProcessingMode::Send);
-
-        // Compute expected values to assert over modified request
-        let mut expected_headers = mock.orig_headers_map();
-        let mut expected_body = mock.body.map(Into::into);
-        let mut expected_trailers = mock.orig_trailers_map();
-        for response_opt in &mock_state.responses {
-            if let Some(response) = response_opt.response.as_ref() {
-                match response {
-                    ProcessingResponseType::RequestHeaders(HeadersResponse { response: Some(common_response) })
-                    | ProcessingResponseType::ResponseHeaders(HeadersResponse { response: Some(common_response) }) => {
-                        let mutation =
-                            should_send_headers.then_some(common_response.header_mutation.as_ref()).flatten();
-                        expected_headers = mock.apply_header_mutation(mutation);
-                    },
-                    ProcessingResponseType::RequestBody(BodyResponse {
-                        response: Some(CommonResponse { body_mutation: Some(BodyMutation { mutation }), .. }),
-                    })
-                    | ProcessingResponseType::ResponseBody(BodyResponse {
-                        response: Some(CommonResponse { body_mutation: Some(BodyMutation { mutation }), .. }),
-                    }) => {
-                        let mutation = should_send_body.then_some(mutation.as_ref()).flatten();
-                        expected_body = mock.apply_body_mutation(mutation);
-                    },
-                    ProcessingResponseType::RequestTrailers(TrailersResponse { header_mutation })
-                    | ProcessingResponseType::ResponseTrailers(TrailersResponse { header_mutation }) => {
-                        let mutation = should_send_trailers.then_some(header_mutation.as_ref()).flatten();
-                        expected_trailers = mock.apply_trailer_mutation(mutation);
-                    },
-                    _ => (),
-                }
-            }
-        }
-
-        assert_eq!(
-            *headers,
-            expected_headers,
-            "test_case #: {}, asserting headers, original {} {:#?}, ext_proc server conf: {:#?}, ext_proc_filter processing mode: {:#?}",
-            test_case_num,
-            <M as MsgKind>::NAME,
-            mock,
-            mock_state,
-            processing_mode
-        );
-        assert_eq!(
-            body,
-            expected_body.as_ref(),
-            "test_case #: {}, asserting body, original {} {:#?}, ext_proc server conf: {:#?}, ext_proc_filter processing mode: {:#?}",
-            test_case_num,
-            <M as MsgKind>::NAME,
-            mock,
-            mock_state,
-            processing_mode
-        );
-        assert_eq!(
-            *trailers,
-            expected_trailers,
-            "test_case #: {}, asserting trailers, original {} {:#?}, ext_proc server conf: {:#?}, ext_proc_filter processing mode: {:#?}",
-            test_case_num,
-            <M as MsgKind>::NAME,
-            mock,
-            mock_state,
-            processing_mode
-        );
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_combinatorial_modes_continue() {
-        let status = ResponseStatus::Continue;
-        let mut test_case_num = 0;
-
-        let mock_states = generate_mock_external_processors_states::<RequestMsg>(status);
-        let processing_modes = generate_processing_mode_configurations::<RequestMsg>();
-        let mock_requests = generate_mock_messages::<RequestMsg>();
-
-        for mock_state in &mock_states {
-            for processing_mode in &processing_modes {
-                for mock_request in &mock_requests {
-                    if validate_mock_server_configuration(mock_request, processing_mode, mock_state) {
-                        debug!(target: "ext_proc_tests", "Test case #: {test_case_num}");
-                        let (server_addr, server_handle) = start_mock_server(mock_state.clone()).await;
-                        let mut config =
-                            create_default_config_for_ext_proc_filter(server_addr, processing_mode.clone());
-                        config.observability_mode = false;
-                        config.failure_mode_allow = false;
-                        let mut ext_proc = ExternalProcessor::from(config);
-                        let mut request = build_request_from_mock(mock_request);
-                        let result = ext_proc.apply_request(&mut request).await;
-                        server_handle.abort();
-                        assert_matches!(result, FilterDecision::Continue);
-
-                        let (parts, body) = request.into_parts();
-                        let request_headers = parts.headers;
-                        let collected = body.collect().await.unwrap();
-                        let request_trailers = if let Some(trailers) = collected.trailers() {
-                            trailers.clone()
-                        } else {
-                            http::HeaderMap::default()
-                        };
-                        let request_body = match collected.to_bytes() {
-                            b if b.is_empty() => None,
-                            b => Some(b),
-                        };
-
-                        assert_result(
-                            test_case_num,
-                            mock_request,
-                            &request_headers,
-                            request_body.as_ref(),
-                            &request_trailers,
-                            mock_state,
-                            processing_mode,
-                        );
-                        test_case_num += 1;
-                    }
-                }
-            }
-        }
-        println!("Total test cases executed: {test_case_num}");
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_response_combinatorial_modes_continue() {
-        let status = ResponseStatus::Continue;
-        let mut test_case_num = 0;
-
-        let mock_states = generate_mock_external_processors_states::<ResponseMsg>(status);
-        let processing_modes = generate_processing_mode_configurations::<ResponseMsg>();
-        let mock_responses = generate_mock_messages::<ResponseMsg>();
-
-        for mock_state in &mock_states {
-            for processing_mode in &processing_modes {
-                for mock_response in &mock_responses {
-                    if validate_mock_server_configuration(mock_response, processing_mode, mock_state) {
-                        debug!(target: "ext_proc_tests", "Test case #: {test_case_num}");
-                        let (server_addr, server_handle) = start_mock_server(mock_state.clone()).await;
-                        let mut config =
-                            create_default_config_for_ext_proc_filter(server_addr, processing_mode.clone());
-                        config.observability_mode = false;
-                        config.failure_mode_allow = false;
-                        let mut ext_proc = ExternalProcessor::from(config);
-                        let mut response = build_response_from_mock(mock_response);
-                        let result = ext_proc.apply_response(&mut response).await;
-                        server_handle.abort();
-                        assert_matches!(result, FilterDecision::Continue);
-
-                        let (parts, body) = response.into_parts();
-                        let response_headers = parts.headers;
-                        let collected = body.collect().await.unwrap();
-                        let response_trailers = if let Some(trailers) = collected.trailers() {
-                            trailers.clone()
-                        } else {
-                            http::HeaderMap::default()
-                        };
-                        let response_body = match collected.to_bytes() {
-                            b if b.is_empty() => None,
-                            b => Some(b),
-                        };
-
-                        assert_result(
-                            test_case_num,
-                            mock_response,
-                            &response_headers,
-                            response_body.as_ref(),
-                            &response_trailers,
-                            mock_state,
-                            processing_mode,
-                        );
-                        test_case_num += 1;
-                    }
-                }
-            }
-        }
-        println!("Total test cases executed: {test_case_num}");
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_combinatorial_modes_observability() {
-        let mut test_case_num = 0;
-
-        let mock_state = MockExternalProcessorState::new();
-        let processing_modes = generate_processing_mode_configurations::<RequestMsg>();
-        let mock_requests = generate_mock_messages::<RequestMsg>();
-
-        for processing_mode in &processing_modes {
-            for mock_request in &mock_requests {
-                if validate_mock_server_configuration(mock_request, processing_mode, &mock_state) {
-                    debug!(target: "ext_proc_tests", "Test case #: {test_case_num}");
-                    let (server_addr, server_handle) = start_mock_server(mock_state.clone()).await;
-                    let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode.clone());
-                    config.observability_mode = true;
-                    config.failure_mode_allow = false;
-                    let mut ext_proc = ExternalProcessor::from(config);
-                    let mut request = build_request_from_mock(mock_request);
-                    let result = ext_proc.apply_request(&mut request).await;
-                    server_handle.abort();
-                    assert_matches!(result, FilterDecision::Continue);
-
-                    let (parts, body) = request.into_parts();
-                    let request_headers = parts.headers;
-                    let collected = body.collect().await.unwrap();
-                    let request_trailers = if let Some(trailers) = collected.trailers() {
-                        trailers.clone()
-                    } else {
-                        http::HeaderMap::default()
-                    };
-                    let request_body = match collected.to_bytes() {
-                        b if b.is_empty() => None,
-                        b => Some(b),
-                    };
-
-                    assert_result(
-                        test_case_num,
-                        mock_request,
-                        &request_headers,
-                        request_body.as_ref(),
-                        &request_trailers,
-                        &mock_state,
-                        processing_mode,
-                    );
-                    test_case_num += 1;
-                }
-            }
-        }
-        println!("Total test cases executed: {test_case_num}");
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_response_combinatorial_modes_observability() {
-        let mut test_case_num = 0;
-
-        let mock_state = MockExternalProcessorState::new();
-        let processing_modes = generate_processing_mode_configurations::<ResponseMsg>();
-        let mock_responses = generate_mock_messages::<ResponseMsg>();
-
-        for processing_mode in &processing_modes {
-            for mock_response in &mock_responses {
-                if validate_mock_server_configuration(mock_response, processing_mode, &mock_state) {
-                    debug!(target: "ext_proc_tests", "Test case #: {test_case_num}");
-                    let (server_addr, server_handle) = start_mock_server(mock_state.clone()).await;
-                    let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode.clone());
-                    config.observability_mode = true;
-                    config.failure_mode_allow = false;
-                    let mut ext_proc = ExternalProcessor::from(config);
-                    let mut response = build_response_from_mock(mock_response);
-                    let result = ext_proc.apply_response(&mut response).await;
-                    server_handle.abort();
-                    assert_matches!(result, FilterDecision::Continue);
-
-                    let (parts, body) = response.into_parts();
-                    let response_headers = parts.headers;
-                    let collected = body.collect().await.unwrap();
-                    let response_trailers = if let Some(trailers) = collected.trailers() {
-                        trailers.clone()
-                    } else {
-                        http::HeaderMap::default()
-                    };
-                    let response_body = match collected.to_bytes() {
-                        b if b.is_empty() => None,
-                        b => Some(b),
-                    };
-
-                    assert_result(
-                        test_case_num,
-                        mock_response,
-                        &response_headers,
-                        response_body.as_ref(),
-                        &response_trailers,
-                        &mock_state,
-                        processing_mode,
-                    );
-                    test_case_num += 1;
-                }
-            }
-        }
-        println!("Total test cases executed: {test_case_num}");
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_header_mutation() {
-        let mock_state = MockExternalProcessorState::new().add_response(create_headers_response::<RequestMsg>(
-            vec![Some(("x-processed", "true")), Some(("x-custom-header", "custom-value"))],
-            None,
-            vec![],
-            ResponseStatus::Continue as i32,
-            None,
-        ));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = false;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request = build_request_from_mock(&Mock::<RequestMsg> {
-            headers: vec![Some(("content-type", "application/json"))],
-            body: None,
-            trailers: vec![],
-            _marker: std::marker::PhantomData,
-        });
-        let result = ext_proc.apply_request(&mut request).await;
-
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(request.headers().get("x-processed").unwrap(), "true");
-        assert_eq!(request.headers().get("x-custom-header").unwrap(), "custom-value");
-        assert_eq!(request.headers().get("content-type").unwrap(), "application/json");
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_header_mutation_pseudo_headers() {
-        let mock_state = MockExternalProcessorState::new().add_response(create_headers_response::<RequestMsg>(
-            vec![
-                Some((":method", "POST")),
-                Some((":path", "/ext-proc-html-path")),
-                Some((":scheme", "https")),
-                Some((":authority", "ext-proc.com")),
-            ],
-            None,
-            vec![],
-            ResponseStatus::Continue as i32,
-            None,
-        ));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode.clone());
-        config.observability_mode = false;
-        config.failure_mode_allow = false;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request = build_request_from_mock(&Mock::<RequestMsg> {
-            headers: vec![Some(("content-type", "application/json"))],
-            body: None,
-            trailers: vec![],
-            _marker: std::marker::PhantomData,
-        });
-        let result = ext_proc.apply_request(&mut request).await;
-        let (parts, _) = request.into_parts();
-
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(parts.method.as_str(), "POST");
-        assert_eq!(parts.uri.authority().unwrap().as_str(), "ext-proc.com");
-        assert_eq!(parts.uri.scheme().unwrap().as_str(), "https");
-        assert_eq!(parts.uri.path_and_query().unwrap().as_str(), "/ext-proc-html-path");
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_trailer_mutation() {
-        let mock_state = MockExternalProcessorState::new()
-            .add_response(create_headers_response::<RequestMsg>(
-                vec![],
-                None,
-                vec![],
-                ResponseStatus::Continue as i32,
-                None,
-            ))
-            .add_response(create_body_response::<RequestMsg>(
-                vec![],
-                None,
-                vec![],
-                ResponseStatus::Continue as i32,
-                None,
-            ))
-            .add_response(create_trailers_response::<RequestMsg>(vec![
-                Some(("x-processed", "true")),
-                Some(("x-custom-trailer", "modified-value")),
-            ]));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::Buffered,
-            request_trailer_mode: TrailerProcessingMode::Send,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = false;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request = build_request_from_mock(&Mock::<RequestMsg> {
-            headers: vec![Some(("content-type", "application/json"))],
-            body: Some("body"),
-            trailers: vec![Some(("x-custom-trailer", "original-value"))],
-            _marker: std::marker::PhantomData,
-        });
-
-        let result = ext_proc.apply_request(&mut request).await;
-        let body = std::mem::take(&mut request.body_mut().inner).collect().await.unwrap();
-        let trailers = body.trailers().cloned();
-
-        assert!(trailers.is_some());
-        let trailers = trailers.unwrap();
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(request.headers().get("content-type").unwrap(), "application/json");
-        assert_eq!(body.to_bytes(), "body");
-        assert_eq!(trailers.get("x-processed").unwrap(), "true");
-        assert_eq!(trailers.get("x-custom-trailer").unwrap(), "modified-value");
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_body_buffered_continue_and_replace_on_headers_response() {
-        let new_body = "modified body content";
-        let mock_state = MockExternalProcessorState::new().add_response(create_headers_response::<RequestMsg>(
-            vec![Some(("y-custom-header", "true"))],
-            Some(new_body.as_bytes().into()),
-            vec![],
-            ResponseStatus::ContinueAndReplace as i32,
-            None,
-        ));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::Buffered,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = false;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request = build_request_from_mock(&Mock::<RequestMsg> {
-            headers: vec![],
-            body: Some("original body"),
-            trailers: vec![],
-            _marker: std::marker::PhantomData,
-        });
-        let result = ext_proc.apply_request(&mut request).await;
-
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(request.method(), Method::GET);
-        assert_eq!(request.headers().get("y-custom-header").unwrap(), "true");
-        let body_bytes = std::mem::take(&mut request.body_mut().inner).collect().await.unwrap().to_bytes();
-        assert_eq!(body_bytes, new_body.as_bytes());
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_body_buffered_continue_and_replace_on_body_response() {
-        let new_body = "modified body content";
-        let mock_state = MockExternalProcessorState::new()
-            .add_response(create_headers_response::<RequestMsg>(
-                vec![],
-                None,
-                vec![],
-                ResponseStatus::Continue as i32,
-                None,
-            ))
-            .add_response(create_body_response::<RequestMsg>(
-                // TODO support headers modifications on body responses
-                //vec![("y-custom-header", "true")],
-                vec![],
-                Some(new_body.as_bytes().into()),
-                vec![],
-                ResponseStatus::ContinueAndReplace as i32,
-                None,
-            ));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::Buffered,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = false;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request = build_request_from_mock(&Mock::<RequestMsg> {
-            headers: vec![],
-            body: Some("original body"),
-            trailers: vec![],
-            _marker: std::marker::PhantomData,
-        });
-        let result = ext_proc.apply_request(&mut request).await;
-
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(request.method(), Method::GET);
-        //assert_eq!(request.headers().get("y-custom-header").unwrap(), "true");
-        let body_bytes = std::mem::take(&mut request.body_mut().inner).collect().await.unwrap().to_bytes();
-        assert_eq!(body_bytes, new_body.as_bytes());
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_body_buffered_mode() {
-        let mock_state = MockExternalProcessorState::new()
-            .add_response(create_headers_response::<RequestMsg>(
-                vec![Some(("x-stream-processed", "true")), Some(("y-custom-header", "true"))],
-                None,
-                vec![],
-                ResponseStatus::Continue as i32,
-                None,
-            ))
-            .add_response(create_body_response::<RequestMsg>(
-                vec![],
-                Some("body data from external processor".as_bytes().into()),
-                vec![],
-                ResponseStatus::Continue as i32,
-                None,
-            ));
-        // starting mock with delay to compare output with the test that sets
-        // send_body_without_waiting_for_header_response to true
-        let (server_addr, _) = start_mock_server_with_delay(mock_state, Duration::from_millis(1)).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::Buffered,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = false;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request = build_request_from_mock(&Mock::<RequestMsg> {
-            headers: vec![],
-            body: Some("buffered body data"),
-            trailers: vec![],
-            _marker: std::marker::PhantomData,
-        });
-        let result = ext_proc.apply_request(&mut request).await;
-
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(request.headers().get("x-stream-processed").unwrap(), "true");
-        let body_bytes = std::mem::take(&mut request.body_mut().inner).collect().await.unwrap().to_bytes();
-        assert_eq!(body_bytes, "body data from external processor".as_bytes());
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_body_buffered_mode_send_body_without_waiting_for_header_response() {
-        let mock_state = MockExternalProcessorState::new()
-            .add_response(create_headers_response::<RequestMsg>(
-                vec![Some(("x-stream-processed", "true")), Some(("y-custom-header", "true"))],
-                None,
-                vec![],
-                ResponseStatus::Continue as i32,
-                None,
-            ))
-            .add_response(create_body_response::<RequestMsg>(
-                vec![],
-                Some("body data from external processor".as_bytes().into()),
-                vec![],
-                ResponseStatus::Continue as i32,
-                None,
-            ));
-        let (server_addr, _) = start_mock_server_with_delay(mock_state, Duration::from_millis(1)).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::Buffered,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = false;
-        config.send_body_without_waiting_for_header_response = true;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request = build_request_from_mock(&Mock::<RequestMsg> {
-            headers: vec![],
-            body: Some("buffered body data"),
-            trailers: vec![],
-            _marker: std::marker::PhantomData,
-        });
-        let result = ext_proc.apply_request(&mut request).await;
-
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(request.headers().get("x-stream-processed").unwrap(), "true");
-        let body_bytes = std::mem::take(&mut request.body_mut().inner).collect().await.unwrap().to_bytes();
-        assert_eq!(body_bytes, "body data from external processor".as_bytes());
-    }
-
-    #[tokio::test]
-    async fn test_request_body_streaming_mode() {
-        let mock_state = MockExternalProcessorState::new()
-            .add_response(create_headers_response::<RequestMsg>(
-                vec![Some(("x-stream-processed", "true")), Some(("y-custom-header", "true"))],
-                None,
-                vec![],
-                ResponseStatus::Continue as i32,
-                // even though we will stream the body, no body modification is
-                // performed in the headers response so no need to set the flag
-                None,
-            ))
-            .add_response(create_body_response::<RequestMsg>(
-                vec![],
-                Some("body data from external processor".as_bytes().into()),
-                vec![],
-                ResponseStatus::Continue as i32,
-                // the response for the streaming body is still just a normal Body
-                // no need to set end_of_stream, which is only for FULL_DUPLEX_STREAMED
-                None,
-            ));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::Streamed,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = false;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request = build_request_from_mock(&Mock::<RequestMsg> {
-            headers: vec![],
-            body: Some("streaming body data"),
-            trailers: vec![],
-            _marker: std::marker::PhantomData,
-        });
-        let result = ext_proc.apply_request(&mut request).await;
-
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(request.headers().get("x-stream-processed").unwrap(), "true");
-        let body_bytes = std::mem::take(&mut request.body_mut().inner).collect().await.unwrap().to_bytes();
-        assert_eq!(body_bytes, "body data from external processor".as_bytes());
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_response_header_mutation_pseudo_headers() {
-        let mock_state = MockExternalProcessorState::new().add_response(create_headers_response::<ResponseMsg>(
-            vec![Some((":status", "404"))],
-            None,
-            vec![],
-            ResponseStatus::Continue as i32,
-            None,
-        ));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Skip,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Send,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = false;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut response = build_response_from_mock(&Mock::<ResponseMsg> {
-            headers: vec![Some(("content-type", "application/json"))],
-            body: None,
-            trailers: vec![],
-            _marker: std::marker::PhantomData,
-        });
-        let result = ext_proc.apply_response(&mut response).await;
-        let (parts, _) = response.into_parts();
-
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(parts.status.as_str(), "404");
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_response_body_streaming_mode() {
-        let mock_state = MockExternalProcessorState::new()
-            .add_response(create_headers_response::<ResponseMsg>(
-                vec![Some(("x-stream-processed", "true")), Some(("y-custom-header", "true"))],
-                None,
-                vec![],
-                ResponseStatus::Continue as i32,
-                // even though we will stream the body, no body modification is
-                // performed in the headers response so no need to set the flag
-                None,
-            ))
-            .add_response(create_body_response::<ResponseMsg>(
-                vec![],
-                Some("body data from external processor".as_bytes().into()),
-                vec![],
-                ResponseStatus::Continue as i32,
-                // the response for the streaming body is still just a normal Body
-                // no need to set end_of_stream, which is only for FULL_DUPLEX_STREAMED
-                None,
-            ));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Skip,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Send,
-            response_body_mode: BodyProcessingMode::Streamed,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = false;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut response =
-            build_response_from_mock(&Mock::<ResponseMsg>::new(vec![], Some("streaming body data"), vec![]));
-        let result = ext_proc.apply_response(&mut response).await;
-
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(response.headers().get("x-stream-processed").unwrap(), "true");
-        let body_bytes = &mut response.body_mut().collect().await.unwrap().to_bytes();
-        assert_eq!(body_bytes, "body data from external processor".as_bytes());
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_response_body_streaming_mode_observability() {
-        let mock_state = MockExternalProcessorState::new();
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Skip,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Send,
-            response_body_mode: BodyProcessingMode::Streamed,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = true;
-        config.failure_mode_allow = false;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut response =
-            build_response_from_mock(&Mock::<ResponseMsg>::new(vec![], Some("streaming body data"), vec![]));
-        let result = ext_proc.apply_response(&mut response).await;
-
-        assert_matches!(result, FilterDecision::Continue);
-        let body_bytes = &mut response.body_mut().collect().await.unwrap().to_bytes();
-        assert_eq!(body_bytes, "streaming body data".as_bytes());
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_header_timeout() {
-        let mock_state = MockExternalProcessorState::new();
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = false;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request =
-            build_request_from_mock(&Mock::<RequestMsg>::new(vec![], Some("streaming body data"), vec![]));
-        let result = ext_proc.apply_request(&mut request).await;
-        assert_matches!(result, FilterDecision::DirectResponse(dr) => {
-            assert_eq!(dr.status(), http::StatusCode::GATEWAY_TIMEOUT);
-        });
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_header_timeout_failure_mode_allow_true() {
-        let mock_state = MockExternalProcessorState::new();
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = true;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request =
-            build_request_from_mock(&Mock::<RequestMsg>::new(vec![], Some("streaming body data"), vec![]));
-        let result = ext_proc.apply_request(&mut request).await;
-        assert_matches!(result, FilterDecision::Continue);
-
-        let (_, body) = request.into_parts();
-        let body_bytes = body.inner.collect().await;
-
-        assert!(body_bytes.is_ok());
-        if let Ok(bytes) = body_bytes {
-            assert_eq!(bytes.to_bytes(), "streaming body data".as_bytes());
-        }
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_body_timeout_failure_mode_allow_false() {
-        let mock_state = MockExternalProcessorState::new();
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Skip,
-            request_body_mode: BodyProcessingMode::Streamed,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = false;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request =
-            build_request_from_mock(&Mock::<RequestMsg>::new(vec![], Some("streaming body data"), vec![]));
-        let result = ext_proc.apply_request(&mut request).await;
-        assert_matches!(result, FilterDecision::Continue);
-
-        let body_bytes = &mut request.body_mut().collect().await;
-        assert!(body_bytes.is_err());
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_body_timeout_failure_mode_allow_true() {
-        let mock_state = MockExternalProcessorState::new();
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Skip,
-            request_body_mode: BodyProcessingMode::Streamed,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = true;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request =
-            build_request_from_mock(&Mock::<RequestMsg>::new(vec![], Some("streaming body data"), vec![]));
-        let result = ext_proc.apply_request(&mut request).await;
-        assert_matches!(result, FilterDecision::Continue);
-
-        let (_, body) = request.into_parts();
-        let body_bytes = body.inner.collect().await;
-
-        assert!(body_bytes.is_ok());
-        if let Ok(bytes) = body_bytes {
-            assert_eq!(bytes.to_bytes(), "streaming body data".as_bytes());
-        }
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_response_header_timeout() {
-        let mock_state = MockExternalProcessorState::new();
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Skip,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Send,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = false;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut response =
-            build_response_from_mock(&Mock::<ResponseMsg>::new(vec![], Some("streaming body data"), vec![]));
-        let result = ext_proc.apply_response(&mut response).await;
-        assert_matches!(result, FilterDecision::DirectResponse(dr) => {
-            assert_eq!(dr.status(), http::StatusCode::GATEWAY_TIMEOUT);
-        });
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_response_header_timeout_failure_mode_allow_true() {
-        let mock_state = MockExternalProcessorState::new();
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Skip,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Send,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = true;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut response =
-            build_response_from_mock(&Mock::<ResponseMsg>::new(vec![], Some("streaming body data"), vec![]));
-        let result = ext_proc.apply_response(&mut response).await;
-        assert_matches!(result, FilterDecision::Continue);
-
-        let (_, body) = response.into_parts();
-        let body_bytes = body.collect().await;
-
-        assert!(body_bytes.is_ok());
-        if let Ok(bytes) = body_bytes {
-            assert_eq!(bytes.to_bytes(), "streaming body data".as_bytes());
-        }
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_response_body_timeout_failure_mode_allow_false() {
-        let mock_state = MockExternalProcessorState::new();
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Skip,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::Streamed,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = false;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut response =
-            build_response_from_mock(&Mock::<ResponseMsg>::new(vec![], Some("streaming body data"), vec![]));
-        let result = ext_proc.apply_response(&mut response).await;
-        assert_matches!(result, FilterDecision::Continue);
-
-        let body_bytes = &mut response.body_mut().collect().await;
-        assert!(body_bytes.is_err());
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_response_body_timeout_failure_mode_allow_true() {
-        let mock_state = MockExternalProcessorState::new();
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Skip,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::Streamed,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.observability_mode = false;
-        config.failure_mode_allow = true;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut response =
-            build_response_from_mock(&Mock::<ResponseMsg>::new(vec![], Some("streaming body data"), vec![]));
-        let result = ext_proc.apply_response(&mut response).await;
-        assert_matches!(result, FilterDecision::Continue);
-
-        let (_, body) = response.into_parts();
-        let body_bytes = body.collect().await;
-
-        assert!(body_bytes.is_ok());
-        if let Ok(bytes) = body_bytes {
-            assert_eq!(bytes.to_bytes(), "streaming body data".as_bytes());
-        }
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_immediate_response_request_header() {
-        let mock_state = MockExternalProcessorState::new().add_response(create_immediate_response(
-            vec![Some(("x-immediate-header", "immediate value"))],
-            Some("immediate body".as_bytes().into()),
-            302,
-        ));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.disable_immediate_response = false;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request =
-            build_request_from_mock(&Mock::<RequestMsg>::new(vec![], Some("streaming body data"), vec![]));
-        let result = ext_proc.apply_request(&mut request).await;
-        assert_matches!(result, FilterDecision::DirectResponse(dr) => {
-            assert_eq!(dr.status(), StatusCode::from_u16(302).unwrap());
-            assert_eq!(dr.headers().get("x-immediate-header").unwrap(), "immediate value");
-            let (_, body) = dr.into_parts();
-            let body_bytes = body.collect().await.unwrap().to_bytes();
-            assert_eq!(body_bytes, "immediate body".as_bytes());
-        });
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_immediate_response_request_header_disable_immediate_response() {
-        let mock_state = MockExternalProcessorState::new().add_response(create_immediate_response(
-            vec![Some(("x-immediate-header", "immediate value"))],
-            Some("immediate body".as_bytes().into()),
-            302,
-        ));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.failure_mode_allow = false;
-        config.disable_immediate_response = true;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request = build_request_from_mock(&Mock::<RequestMsg>::new(
-            vec![Some(("x-test-header", "original value"))],
-            Some("streaming body data"),
-            vec![],
-        ));
-        let result = ext_proc.apply_request(&mut request).await;
-
-        assert_matches!(result, FilterDecision::DirectResponse(dr) => {
-            assert_eq!(dr.status(), StatusCode::from_u16(500).unwrap());
-        });
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_immediate_response_request_header_disable_immediate_response_failure_allow_mode() {
-        let mock_state = MockExternalProcessorState::new().add_response(create_immediate_response(
-            vec![Some(("x-immediate-header", "immediate value"))],
-            Some("immediate body".as_bytes().into()),
-            302,
-        ));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.failure_mode_allow = true;
-        config.disable_immediate_response = true;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request = build_request_from_mock(&Mock::<RequestMsg>::new(
-            vec![Some(("x-test-header", "original value"))],
-            Some("streaming body data"),
-            vec![],
-        ));
-        let result = ext_proc.apply_request(&mut request).await;
-
-        assert_matches!(result, FilterDecision::Continue);
-        let (parts, body) = request.into_parts();
-        assert_eq!(parts.headers.get("x-test-header").unwrap(), "original value");
-        let body_bytes = body.collect().await;
-        assert!(body_bytes.is_ok());
-        if let Ok(bytes) = body_bytes {
-            assert_eq!(bytes.to_bytes(), "streaming body data".as_bytes());
-        }
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_header_override_request_body() {
-        // Original processing mode
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        // ext-proc override
-        let mut header_response =
-            create_headers_response::<RequestMsg>(vec![], None, vec![], ResponseStatus::Continue as i32, None);
-        header_response.mode_override = Some(EnvoyProcessingMode {
-            request_header_mode: processing_mode::HeaderSendMode::Send as i32,
-            request_body_mode: processing_mode::BodySendMode::Buffered as i32,
-            request_trailer_mode: processing_mode::HeaderSendMode::Skip as i32,
-            response_header_mode: processing_mode::HeaderSendMode::Skip as i32,
-            response_body_mode: processing_mode::BodySendMode::None as i32,
-            response_trailer_mode: processing_mode::HeaderSendMode::Skip as i32,
-        });
-        let mock_state =
-            MockExternalProcessorState::new().add_response(header_response).add_response(create_body_response::<
-                RequestMsg,
-            >(
-                vec![],
-                Some("ext-proc body".into()),
-                vec![],
-                ResponseStatus::Continue as i32,
-                None,
-            ));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.allow_mode_override = true;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request = build_request_from_mock(&Mock::<RequestMsg> {
-            headers: vec![Some(("content-type", "application/json"))],
-            body: Some("original body"),
-            trailers: vec![],
-            _marker: std::marker::PhantomData,
-        });
-
-        let result = ext_proc.apply_request(&mut request).await;
-        assert_matches!(result, FilterDecision::Continue);
-
-        assert_eq!(request.headers().get("content-type").unwrap(), "application/json");
-
-        let (_, body) = request.into_parts();
-        let body_bytes = body.collect().await;
-        assert!(body_bytes.is_ok());
-        if let Ok(bytes) = body_bytes {
-            assert_eq!(bytes.to_bytes(), "ext-proc body".as_bytes());
-        }
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_header_override_request_trailers() {
-        // Original processing mode
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        // ext-proc override
-        let mut header_response =
-            create_headers_response::<RequestMsg>(vec![], None, vec![], ResponseStatus::Continue as i32, None);
-        header_response.mode_override = Some(EnvoyProcessingMode {
-            request_header_mode: processing_mode::HeaderSendMode::Send as i32,
-            request_body_mode: processing_mode::BodySendMode::None as i32,
-            request_trailer_mode: processing_mode::HeaderSendMode::Send as i32,
-            response_header_mode: processing_mode::HeaderSendMode::Skip as i32,
-            response_body_mode: processing_mode::BodySendMode::None as i32,
-            response_trailer_mode: processing_mode::HeaderSendMode::Skip as i32,
-        });
-        let mock_state =
-            MockExternalProcessorState::new().add_response(header_response).add_response(create_trailers_response::<
-                RequestMsg,
-            >(vec![Some((
-                "x-ext-proc-trailer",
-                "ext-proc trailer value",
-            ))]));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.allow_mode_override = true;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request = build_request_from_mock(&Mock::<RequestMsg> {
-            headers: vec![Some(("content-type", "application/json"))],
-            body: None,
-            trailers: vec![Some(("x-original-trailer", "original trailer value"))],
-            _marker: std::marker::PhantomData,
-        });
-
-        let result = ext_proc.apply_request(&mut request).await;
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(request.headers().get("content-type").unwrap(), "application/json");
-
-        let body = std::mem::take(&mut request.body_mut().inner).collect().await.unwrap();
-        let trailers = body.trailers().cloned();
-
-        assert!(trailers.is_some());
-        let trailers = trailers.unwrap();
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(trailers.get("x-ext-proc-trailer").unwrap(), "ext-proc trailer value");
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_response_header_override_response_body() {
-        // Original processing mode
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Skip,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Send,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        // ext-proc override
-        let mut header_response =
-            create_headers_response::<ResponseMsg>(vec![], None, vec![], ResponseStatus::Continue as i32, None);
-        header_response.mode_override = Some(EnvoyProcessingMode {
-            request_header_mode: processing_mode::HeaderSendMode::Skip as i32,
-            request_body_mode: processing_mode::BodySendMode::None as i32,
-            request_trailer_mode: processing_mode::HeaderSendMode::Skip as i32,
-            response_header_mode: processing_mode::HeaderSendMode::Send as i32,
-            response_body_mode: processing_mode::BodySendMode::Buffered as i32,
-            response_trailer_mode: processing_mode::HeaderSendMode::Skip as i32,
-        });
-        let mock_state =
-            MockExternalProcessorState::new().add_response(header_response).add_response(create_body_response::<
-                ResponseMsg,
-            >(
-                vec![],
-                Some("ext-proc body".into()),
-                vec![],
-                ResponseStatus::Continue as i32,
-                None,
-            ));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.allow_mode_override = true;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut response = build_response_from_mock(&Mock::<ResponseMsg> {
-            headers: vec![Some(("content-type", "application/json"))],
-            body: Some("original body"),
-            trailers: vec![],
-            _marker: std::marker::PhantomData,
-        });
-
-        let result = ext_proc.apply_response(&mut response).await;
-        assert_matches!(result, FilterDecision::Continue);
-
-        assert_eq!(response.headers().get("content-type").unwrap(), "application/json");
-        let (_, body) = response.into_parts();
-        let body_bytes = body.collect().await;
-        assert!(body_bytes.is_ok());
-        if let Ok(bytes) = body_bytes {
-            assert_eq!(bytes.to_bytes(), "ext-proc body".as_bytes());
-        }
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_response_header_override_response_trailers() {
-        // Original processing mode
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Skip,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Send,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        // ext-proc override
-        let mut header_response =
-            create_headers_response::<ResponseMsg>(vec![], None, vec![], ResponseStatus::Continue as i32, None);
-        header_response.mode_override = Some(EnvoyProcessingMode {
-            request_header_mode: processing_mode::HeaderSendMode::Skip as i32,
-            request_body_mode: processing_mode::BodySendMode::None as i32,
-            request_trailer_mode: processing_mode::HeaderSendMode::Skip as i32,
-            response_header_mode: processing_mode::HeaderSendMode::Send as i32,
-            response_body_mode: processing_mode::BodySendMode::None as i32,
-            response_trailer_mode: processing_mode::HeaderSendMode::Send as i32,
-        });
-        let mock_state =
-            MockExternalProcessorState::new().add_response(header_response).add_response(create_trailers_response::<
-                ResponseMsg,
-            >(vec![Some((
-                "x-ext-proc-trailer",
-                "ext-proc trailer value",
-            ))]));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.allow_mode_override = true;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut response = build_response_from_mock(&Mock::<ResponseMsg> {
-            headers: vec![Some(("content-type", "application/json"))],
-            body: None,
-            trailers: vec![Some(("x-original-trailer", "original trailer value"))],
-            _marker: std::marker::PhantomData,
-        });
-
-        let result = ext_proc.apply_response(&mut response).await;
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(response.headers().get("content-type").unwrap(), "application/json");
-
-        let (_, body) = response.into_parts();
-        let collected_body = body.collect().await.unwrap();
-        let trailers = collected_body.trailers().cloned();
-        assert!(trailers.is_some());
-        let trailers = trailers.unwrap();
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(trailers.get("x-ext-proc-trailer").unwrap(), "ext-proc trailer value");
-    }
-
-    #[tokio::test]
-    #[test_log::test]
-    async fn test_request_header_override_response_header() {
-        // Original processing mode
-        let processing_mode = ProcessingMode {
-            request_header_mode: HeaderProcessingMode::Send,
-            request_body_mode: BodyProcessingMode::None,
-            request_trailer_mode: TrailerProcessingMode::Skip,
-            response_header_mode: HeaderProcessingMode::Skip,
-            response_body_mode: BodyProcessingMode::None,
-            response_trailer_mode: TrailerProcessingMode::Skip,
-        };
-
-        // ext-proc override
-        let mut request_header_response =
-            create_headers_response::<RequestMsg>(vec![], None, vec![], ResponseStatus::Continue as i32, None);
-        request_header_response.mode_override = Some(EnvoyProcessingMode {
-            request_header_mode: processing_mode::HeaderSendMode::Send as i32,
-            request_body_mode: processing_mode::BodySendMode::None as i32,
-            request_trailer_mode: processing_mode::HeaderSendMode::Skip as i32,
-            response_header_mode: processing_mode::HeaderSendMode::Send as i32,
-            response_body_mode: processing_mode::BodySendMode::Buffered as i32,
-            response_trailer_mode: processing_mode::HeaderSendMode::Send as i32,
-        });
-        let mock_state = MockExternalProcessorState::new()
-            .add_response(request_header_response)
-            .add_response(create_headers_response::<ResponseMsg>(
-                vec![Some(("x-custom-header", "ext-proc header value"))],
-                None,
-                vec![],
-                ResponseStatus::Continue as i32,
-                None,
-            ))
-            .add_response(create_body_response::<ResponseMsg>(
-                vec![],
-                Some("ext-proc body value".into()),
-                vec![],
-                ResponseStatus::Continue as i32,
-                None,
-            ))
-            .add_response(create_trailers_response::<ResponseMsg>(vec![Some((
-                "x-custom-trailer",
-                "ext-proc trailer value",
-            ))]));
-        let (server_addr, _) = start_mock_server(mock_state).await;
-
-        let mut config = create_default_config_for_ext_proc_filter(server_addr, processing_mode);
-        config.allow_mode_override = true;
-        let mut ext_proc = ExternalProcessor::from(config);
-
-        let mut request = build_request_from_mock(&Mock::<RequestMsg> {
-            headers: vec![Some(("content-type", "application/json"))],
-            body: None,
-            trailers: vec![],
-            _marker: std::marker::PhantomData,
-        });
-
-        let request_result = ext_proc.apply_request(&mut request).await;
-        assert_matches!(request_result, FilterDecision::Continue);
-        assert_eq!(request.headers().get("content-type").unwrap(), "application/json");
-
-        let mut response = build_response_from_mock(&Mock::<ResponseMsg> {
-            headers: vec![Some(("content-type", "application/json"))],
-            body: Some("original response body"),
-            trailers: vec![Some(("x-custom-trailer", "original trailer value"))],
-            _marker: std::marker::PhantomData,
-        });
-
-        let result = ext_proc.apply_response(&mut response).await;
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(response.headers().get("content-type").unwrap(), "application/json");
-        assert_eq!(response.headers().get("x-custom-header").unwrap(), "ext-proc header value");
-
-        let (_, body) = response.into_parts();
-        let collected_body = body.collect().await.unwrap();
-        let trailers = collected_body.trailers().cloned();
-        assert!(trailers.is_some());
-        assert_eq!(collected_body.to_bytes(), "ext-proc body value");
-        let trailers = trailers.unwrap();
-        assert_matches!(result, FilterDecision::Continue);
-        assert_eq!(trailers.get("x-custom-trailer").unwrap(), "ext-proc trailer value");
     }
 }
